% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfPlot.R
\name{cdfPlot}
\alias{cdfPlot}
\title{Create deluxe CDF Plots}
\usage{
cdfPlot(
  contrastDF,
  pvalCol = "P.Value",
  pThreshold = 0.01,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  insetTitle = NULL,
  symbolSize = c(2, 1),
  symbolShape = c(20, 20),
  symbolColor = c("red3", "deepskyblue4"),
  symbolFill = c("red3", "deepskyblue4"),
  alpha = 1,
  referenceLine = NULL,
  refLineThickness = 1,
  legendPosition = "se",
  baseFontSize = 12,
  viewportX = 0.15,
  viewportY = 0.93,
  viewportWidth = 0.35,
  themeStyle = "grey",
  pvalMax = 0.1,
  printPlot = TRUE,
  footnote,
  footnoteSize = 3,
  footnoteColor = "black",
  footnoteJust = 1
)
}
\arguments{
\item{contrastDF}{A dataframe with LogRatio and LogIntensity columns and optionally a p-value or FDR column.}

\item{pvalCol}{Name of the p-value or FDR column (Default = "P.Value")}

\item{pThreshold}{Used to color points (default = 0.01)}

\item{xlab}{X axis label (default is "LogIntensity column name "Rank")}

\item{ylab}{Y axis label (default is p-value column name)}

\item{title}{Plot title (Optional)}

\item{insetTitle}{Title for the inset plot (Optional)}

\item{symbolSize}{Size of symbols for up, no change, and down. Default = c(4, 1, 4);
Note: All three cannot be the same size. Decimal values are acceptable to help offset that
(e.g. 4, 4.1, 4.2).}

\item{symbolShape}{Shape of the symbols for up, no change, and down; Default = c(21, 20, 21)
(20 = filled circle, 21 = fillable open circle); Note: The same symbol shape cannot
be selected for all three symbols.
See \url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{c(Up, NoChange, Down); default = c("black", "grey25", "grey0")
See \url{http://research.stowers-institute.org/efg/R/Color/Chart}
Note: Colors cannot be duplicated.}

\item{symbolFill}{Set the fill color for the symbols. Note only symbols 21-25 are fillable.
This will have no effect on other symbols.
Default = c("red3", "grey25", "deepskyblue4")
Note: Colors cannot be duplicated.}

\item{alpha}{Controls the transparency of the plotted points (0-1; Default =
0.5)}

\item{referenceLine}{Color for an horizontal line drawn at the p-threshold
(Default = NULL; NULL disables, set to desired color to enable)}

\item{refLineThickness}{Set thickness of the reference line (Default = 1)}

\item{legendPosition}{(Default = "se")}

\item{baseFontSize}{The smallest size font in the figure in points. (Default
= 12)}

\item{viewportX}{(Default = 0.15)}

\item{viewportY}{(Default = 0.93)}

\item{viewportWidth}{(Default = 0.35)}

\item{themeStyle}{"bw" or "grey" which correspond to theme_bw or theme_grey
respectively. (Default = "grey")}

\item{pvalMax}{Limit the range of the main plot (Default = 0.10)}

\item{printPlot}{Specify printing the combined plot to the console/knitr
(Default = TRUE)}

\item{footnote}{Optional string placed right justified at bottom of plot.}

\item{footnoteSize}{Applies to footnote. (Default = 3)}

\item{footnoteColor}{Applies to footnote. (Default = "black")}

\item{footnoteJust}{Value 0 - 1. 0 is left justified, 1 is right justified, 0.5 is centered. (Default = 1)}
}
\value{
A list containing main plot, inset plot, and viewport. The plot can be
   reconstructed with print(myList$main); print(inset, vp = myList$viewport)
}
\description{
CDF plots are a good complement to p-value histograms as a way to evaluate
model performance and examine support for differential expression. Results
are ranked by p-value on the x-axis and the p-value plotted on the y-axis.
Since p-value distributions should be flat, this type of plot should produce a
straight line.  Any observations that fail to meet the null hypothesis will
appear as a break in the line at the low end of the curve.
}
\details{
This function is designed to take topTable dataframes and display the
corresponding CDF plot. Data for the p-values below 0.1 (configurable via
pvalMax argument) are show in a full size plot. An inset figure shows the
whole p-value scale and highlights the portion shown in the full plot. Points
below 0.01 are a different color by default (threshold set by pThreshold
argument; shape/color attributes customizable through other arguments).

The output can be controlled using printPlot = TRUE, which outputs the compound plot
to the console/knitr.

\strong{Data Structure for the input dataframe:}

The defaults are set for dataframes produced by topTable.  The column
"P.Value" is used by default to accommodate the column names used in topTable
dataframes.  Any other dataframe can be used with by explicitly defining the
p-value column name with the appropriate argument.

Sensible defaults are chosen for symbols (Size, Shape, Color, and Fill).
There are optional arguments that allow these to be adjusted. A length of 2
is required for these arguments which applies the attributes in
this order: Significant, Not Significant.
}
\examples{
\dontrun{
   # Plot to console (contrastDF is a topTable dataframe)
   cdfPlot(contrastDF, title = "My CDF Plot")
}
}
