% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlot.R
\name{comparePrep}
\alias{comparePrep}
\title{Create a data.frame to use with comparePlot from topTable data.frames}
\usage{
comparePrep(contrastList, valueCol = "logFC", significanceCol = "P.Value")
}
\arguments{
\item{contrastList}{A named list of 2 topTable dataframes (Required). The
names are used as column names for the value columns in the output.}

\item{valueCol}{Name of column containing values to plot (Default = "logFC")}

\item{significanceCol}{Name of column to use for significance (Default = "P.Value")}
}
\value{
A data frame with 2 LogRatio measurements and 2 significance columns.  Columns 1 and 3 map
to sample 1 and columns 2 and 4 map to sample 2.  The returned dataframe is formatted as expected
by the comparePlot function.
}
\description{
Takes two topTable dataframes and outputs a dataframe suitable for function
comparePlot() (2 columns of LogRatio data and 2 columns of significant
measures). Filter the two topTable to contain only the intersecting genes
(present in both datasets). The two dataframes must have the same type of
gene IDs as rownames.
}
\examples{
\dontrun{
  # Retrieve the 1st two contrasts from a DGEobj
  contrastList <- getType(dgeObj, "topTable")[1:2]

  # Capture the default logFC and P.Value
  compareDat <- comparePrep(contrastList)

  # Switch to an FDR value for the significance measure
  compareDat <- comparePrep(contrastList, significanceCol="adj.P.Val")

  # Draw the plot
  cPlot <- comparePlot(compareDat)
}

}
