% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEobj.plots.R
\docType{package}
\name{DGEobj.plots-package}
\alias{DGEobj.plots-package}
\title{DGEobj.plots Package Overview}
\description{
TBD
}
\section{More Information}{

\code{browseVignettes(package = 'DGEobj.plots')}
}

