% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{MDS_var_explained}
\alias{MDS_var_explained}
\title{Explain variance of MDS object or log2 matrix}
\usage{
MDS_var_explained(
  mds,
  plotType = "canvasXpress",
  topN = 10,
  cumVarLimit = 0.9,
  barColor = "dodgerblue4",
  barWidth = 0.65
)
}
\arguments{
\item{mds}{A class MDS object from limma::plotMDS() or a data matrix to analyze
(typically log2) (required)}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{topN}{The number of dimensions to plot (default = 10)}

\item{cumVarLimit}{The maximum cumulative variance to plot. Range 0-1. (default = 0.9)}

\item{barColor}{Fill and outline color default = "dodgerblue4"}

\item{barWidth.}{Range 0-1. Applicable only for ggplot. (default = 0.65)}
}
\value{
A list with two plots(ggplot or canvasXpress plots) and the variance explained data.frame.
}
\description{
Takes a class MDS object from limma::plotMDS() and generates two plots: 1)
fraction of variance for each dimension, 2) cumulative variance. By default,
it plots the first 10 dimensions or the first N dimensions totaling 90%.
}
\examples{
\dontrun{
     # Plot the first two dimensions
     MyMDS <- ggplotMDS(MyDGEobj)
     MyMDS[[1]]  #the MDS plot

     # Then apply MDS_var_explained (the MDS object is MyMDS[[2]])
     varResults <- MDS_var_explained(MyMDS[[2]])
     varResults[[1]] # The Variance per dimension plot
     varResults[[2]] # The cumulative variance plot
     var_explained <- varResults[[3]]  # Data used for plotting (unfiltered)

     # MDS_var_explained - ggplot
     varResults_ggplot <- MDS_var_explained(MyMDS[[2]], plotType = "ggplot")
     varResults_ggplot[[1]] # The Variance per dimension plot
     varResults_ggplot[[2]] # The cumulative variance plot
     var_explained <- varResults_ggplot[[3]]  # Data used for plotting (unfiltered)
}

}
