% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot.R
\name{obsPlot}
\alias{obsPlot}
\title{Function obsPlot}
\usage{
obsPlot(
  data,
  plotType = "canvasXpress",
  group,
  boxLayer = TRUE,
  violinLayer = FALSE,
  showPoints = TRUE,
  showMean = TRUE,
  xlab,
  ylab,
  title,
  boxColor = "deepskyblue3",
  boxTransparency = 0.5,
  violinColor = "goldenrod1",
  violinTransparency = 0.5,
  facet = TRUE,
  facetRow,
  labelAngle = 30,
  axisFree = TRUE
)
}
\arguments{
\item{data}{A  dataframe row and colnames (required)}

\item{group}{Define the column name to group boxplots by (typically a replicate group column) (required)}

\item{boxLayer}{Adds a boxplot layer (default = TRUE)}

\item{violinLayer}{Adds a violin layer (default = FALSE)}

\item{showPoints}{Shows the datapoints on the plot only if boxLayer or violinLayer is enabled. (default = TRUE)}

\item{showMean}{Show the mean of the points on the plot only if boxLayer is enabled. (default = TRUE)}

\item{xlab}{X axis label (defaults to group column name if not specified)}

\item{ylab}{Y axis label (defaults to value column name if not specified)}

\item{title}{Plot title (optional)}

\item{boxColor}{Color for the boxplot layer (default = "deepskyblue3")}

\item{boxTransparency}{Transparency for the box layer (default = 0.5)}

\item{violinColor}{Color for the violin layer (default = "yellow")}

\item{violinTransparency}{Transparency for the violin layer (default = 0.5)}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE)}

\item{facetRow}{Explicitly set the number of rows for the facet plot.
Default behavior will automatically set the columns. (default = NULL)}

\item{labelAngle}{Angle to set the sample labels on the X axis. (default =  30; range = 0-90)}

\item{axisFree}{Specify same scale or independent scales for each subplot (Default = TRUE;
Allowed values: TRUE and FALSE)}

\item{boxFill}{Fill Color for the boxplot layer (default = "deepskyblue3")}

\item{violinFill}{Fill Color for the violin (default = "yellow")}
}
\value{
Plot of type canvasXpress or ggplot. If Facet = TRUE (default) returns a faceted object. If
  facet = FALSE, returns a list of objects indexed
  by observation (gene) names.
}
\description{
Provides a summary plot (box/violin/points) for each observation (gene), showing data for each
experiment group. The plot can optionally include either one or both of the boxplot and violinplot layer
can additionally choose to display points and/or mean of all the points on the ploit.  The boxLayer and the
violin layer can be customized to have the desired transparency, colors etc. By default, the violinLayer is
not displayed and only the boxplot, points and the mean on the plots can be see.
}
\details{
Also, by default, the plots are faceted.  Faceting the plot can be turned off to return a list of individual
plots for each gene.

Input is a dataframe or matrix of intensity values with geneID row names and sample name column names,
e.g. from log2cpm. Group must also be specified as the plots are grouped by this attribute.
}
\examples{
\dontrun{
  # Simulate some data with row and colnames
  groups <- paste("group", factor(rep(1:4, each = 100)), sep = "")
  x <- matrix(rnorm(2400, mean = 10), ncol = length(groups))
  colnames(x) <- paste("sample", 1:ncol(x), sep = "")
  rownames(x) <- paste("gene", 1:nrow(x), sep="")

 # Or get data from a DGEobj with RNA-Seq data
 log2cpm <- convertCounts(dgeObj$counts, unit = "cpm", log = TRUE, normalize = "tmm")
 design <- getItem(t_obj1,"design")
 replicategroup <- design[,"ReplicateGroup",drop = FALSE]

  # Faceted boxplot
  obsPlot(log2cpm,
          group = replicategroup)

  # Faceted violin plot
  obsPlot(log2cpm,
           violinLayer = TRUE,
           boxLayer = FALSE,
           group = replicategroup)

  # Return a list of plot for each individual gene
  myplots <- obsPlot(log2cpm,
                     group = replicategroup
                     facet = FALSE)
  # Plot one from the list
  myplots[[2]]
}

}
