% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPValHist.R
\name{plotPvalHist}
\alias{plotPvalHist}
\title{Plot histogram analysis of p-value distributions}
\usage{
plotPvalHist(
  P.Val,
  plotType = "canvasXpress",
  facet = TRUE,
  binWidth = 0.02,
  transparency = 0.6,
  color = "dodgerblue3"
)
}
\arguments{
\item{P.Val}{A matrix or dataframe of numeric data; col = samples}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{facet}{Set to FALSE to print individual plots instead of a faceted plot. (default = TRUE)}

\item{binWidth}{Value is always between 0 and 1. (default = 0.02)}

\item{transparency}{Set the transparency. (default = 0.6)}

\item{color}{Fill & Outline color for the histogram (default = "dodgerblue3")}
}
\value{
A canvasXpress or a ggplot2 object if facet = TRUE or a list of plots if facet = FALSE. (default = TRUE)
}
\description{
Generate a facet plot (or optionally individual plots) from a dataframe of
numbers. Intended to perform histogram analysis of p-value distributions,
but should be useful for any dataframe of numeric columns.
}
\examples{
\dontrun{
   # Print to console using all defaults
   MyPvalMatrix <- extractCol(getType(myDGEobj, "topTable"), "P.Value")
   plotPvalHist(MyPvalMatrix)

   # Use some custom arguments
   myplot <- plotPvalHist(MyPValMatrix)
}

}
