% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Plot log intensity versus log ratio}
\usage{
profilePlot(
  dgeObj,
  contrast,
  plotType = "canvasXpress",
  logRatioCol = "logFC",
  logIntCol = "AveExpr",
  pvalCol = "P.Value",
  pthreshold = 0.01,
  geneNameCol,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  sizeBySignificance = FALSE,
  referenceLine = "darkgoldenrod1",
  foldChangeThreshold = 1.5,
  lineFitType = "loess"
)
}
\arguments{
\item{dgeObj}{DGEobj with a topTable itemType (required).}

\item{contrast}{itemName of the contrast in the dgeObj.}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = "canvasXpress").}

\item{logRatioCol}{Name of the LogRatio column (default = "logFC")}

\item{logIntCol}{Name of the LogIntensity column (default = "AveExpr")}

\item{pvalCol}{Name of the p-value column (default = "P.Value")}

\item{pthreshold}{the p-value threshold used to color points (default = 0.01)}

\item{geneNameCol}{geneName column in geneData from DGEobj. This column will be used to add the gene name
in the popover for canvasXpress plots only.}

\item{xlab}{X axis label (defaults = "LogIntensity column name")}

\item{ylab}{Y axis label (defaults = "LogRatio column name")}

\item{title}{Plot title (optional)}

\item{sizeBySignificance}{Set to TRUE to size points by the negative Log10 of the
Significance measure (default = FALSE)}

\item{referenceLine}{Color for a horizontal reference line at intercept = 0
(default = "darkgoldenrod1"; NULL disables)}

\item{foldChangeThreshold}{Position of reference horizontal lines for fold change
(default = 1.5)}

\item{lineFitType}{Enable a line fit through the data (default = "loess";
"lm" produces a linear fit. NULL disables)}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
A profile plot shows Log Intensity on the X axis and Log Ratio on the Y axis.
This function is intended to show the profile plot from a topTable dataframe
in DGEobj (i.e BDL_vs_SHAM).  Properly normalized data will generally be
centered around LogRatio = 0.
}
\details{
By default, the plot places "logFC" on the Y axis and "AveExpr" on the X
axis. By default, a reference horizontal line is shown at LogRatio = 0 on the
Y axis. Optionally, additional reference lines will be drawn at +/- a user
supplied LogRatio threshold. A Loess line fit is drawn through the actual
data. The points are color coded using the significance measure (i.e p-value)
supplied by the user. By default, the P.Value field is used
with a threshold of 0.01 to color code the points.

\strong{Data Structure for the input dataframe:}

The defaults are set for the contrasts on the DGEobj object.
The columns named "logFC", "AveExpr", and "P.Value" are used by default to accommodate
the column names used in these dataframes.  Any other dataframe can be used with fold-change,
intensity, and significance measures, with appropriate arguments to define the column names
to use provided. By default, the column names will be used for the axis labels,
but can be overridden with xlab and ylab arguments.

A significance measure (which defaults to P.Value <= 0.01) is used to color code genes that
are significantly increased or decreased.
}
\examples{
\dontrun{
   # Get DGEObj
   t_obj1 <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
   # Get the name of the contrast data to plot
   contrasts <- names(DGEobj::getType(t_obj1, "topTable"))
   interestedContrast <- contrasts[1]

   # Some options with a custom datafile
   myPlot <- profilePlot(t_obj1,
                         interestedContrast,
                         pthreshold = 0.1,
                         title = "BDL_vs_Sham",
                         referenceLine = "blue")

   myPlot <- profilePlot(t_obj1,
                         interestedContrast,
                         pthreshold = 0.1,
                         title = "BDL_vs_Sham",
                         referenceLine = "blue",
                         plotType = "ggplot")
}

}
