% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfPlot.R
\name{cdfPlot}
\alias{cdfPlot}
\title{Create deluxe CDF Plots}
\usage{
cdfPlot(
  dgeObj,
  contrast,
  plotType = "canvasXpress",
  pvalCol = "P.Value",
  pThreshold = 0.01,
  xlab,
  ylab,
  title = NULL,
  viewportTitle = NULL,
  referenceLine = NULL,
  viewportX = 0.15,
  viewportY = 0.85,
  viewportWidth = 0.35,
  pvalMax = 0.1
)
}
\arguments{
\item{dgeObj}{A DGEobj with one or more topTables (required)}

\item{contrast}{Name of a topTable dataframe with p-value or an FDR column (required)}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{pvalCol}{Name of the p-value or FDR column (default = "P.Value")}

\item{pThreshold}{Used to color points (default = 0.01)}

\item{xlab}{X axis label (default = "Rank")}

\item{ylab}{Y axis label (default = p-value column name)}

\item{title}{Plot title (Optional)}

\item{viewportTitle}{Title for the inset plot (Optional)}

\item{referenceLine}{Color for a horizontal line drawn at the p-threshold
(default = NULL; NULL disables, set to desired color to enable)}

\item{viewportX}{x-location for the inset plot(default = 0.15)}

\item{viewportY}{y-location for the inset plot(default = 0.85)}

\item{viewportWidth}{width of the inset plot (default = 0.35)}

\item{pvalMax}{Limit the range of the main plot (default = 0.10)}
}
\value{
A list containing main plot, inset plot for both plotTypes. For plotType ="ggplot" output list also contains a combined plot which
displays the inset plot in a viewport
}
\description{
CDF plots are a good complement to p-value histograms as a way to evaluate
model performance and examine support for differential expression. Results
are ranked by p-value on the x-axis and the p-value plotted on the y-axis.
Since p-value distributions should be flat, this type of plot should produce a
straight line.  Any observations that fail to meet the null hypothesis will
appear as a break in the line at the low end of the curve.
}
\details{
This function is designed to take a topTable dataframe and display the
corresponding CDF plot. Data for the p-values below 0.1 (configurable via
pvalMax argument) are shown in a full size plot. An viewport figure shows the
whole p-value scale. Points below 0.01 are a different color by default
(threshold set by pThreshold argument; shape/color attributes customizable
through other arguments).

\strong{Data Structure for the input dataframe:}

The defaults are set for dataframes produced by topTable.  The column
"P.Value" is used by default to accommodate the column names used in topTable
dataframes.  Any other dataframe can be used with by explicitly defining the
p-value column named with the appropriate argument.
}
\examples{
\dontrun{
   # Plot to console (dgeObj is a DGEobj and contrast is a name of toptable dataframe from DGEobj)
   contrast <- names(DGEobj::getType(dgeObj, type = "topTable"))

   cdfPlot(dgeObj, contrast = contrast[1], title = "My CDF Plot")

   cdfPlot(dgeObj, contrast = contrast[1], title = "My CDF Plot", plotType = "ggplot")
}
}
