% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfPlot.R
\name{cdfPlot}
\alias{cdfPlot}
\title{Create deluxe CDF Plots}
\usage{
cdfPlot(
  dgeObj,
  contrast,
  plotType = "canvasXpress",
  pvalCol = "P.Value",
  pThreshold = 0.01,
  xlab,
  ylab,
  title = NULL,
  insetTitle = NULL,
  referenceLine = NULL,
  insetX = 0.15,
  insetY = 0.85,
  insetWidth = 0.35,
  pvalMax = 0.1
)
}
\arguments{
\item{dgeObj}{A DGEobj with one or more topTables}

\item{contrast}{Name of a topTable dataframe with p-value or an FDR column}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{pvalCol}{Name of the p-value or FDR column (default = "P.Value")}

\item{pThreshold}{Significant value threshold (default = 0.01)}

\item{xlab}{X axis label (default = "Rank")}

\item{ylab}{Y axis label (default = p-value column name)}

\item{title}{Plot title (Optional)}

\item{insetTitle}{Title for the inset plot, only for ggplots (Optional)}

\item{referenceLine}{Color for a horizontal line drawn at the p-threshold
(default = NULL; NULL disables, set to desired color to enable)}

\item{insetX}{x-location for the inset plot, only for ggplots (default = 0.15)}

\item{insetY}{y-location for the inset plot, only for ggplots (default = 0.85)}

\item{insetWidth}{width of the inset plot, only for ggplots (default = 0.35)}

\item{pvalMax}{Limit the range of the main plot (default = 0.10)}
}
\value{
A list of plots -- 2 plots if plotType is 'canvasXpress' or 3 plots if plotType is 'ggplot'.
}
\description{
CDF plots are a good complement to p-value histograms as a way to evaluate
model performance and examine support for differential expression. On the x-axis,
is the ranked by p-value and on y-axis is the p-value.
Since p-value distributions should be flat, this type of plot should produce a
straight line. Any observations that fail to meet the null hypothesis will
have a break in the line at the end of the curve.
}
\details{
This function is designed to take a topTable dataframe and display the
corresponding CDF plots. The first plot (below_pvalMax) are p-values below the pvalMax and the
the second plot (all_pval) shows all the p-values. For plotType ="ggplot", output also
contains an additional plot that displays the all_pval plot as an inset in below_pvalMax.
}
\examples{
\dontrun{
   dgeObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
   # Plot to console (dgeObj is a DGEobj and contrast is a name of toptable dataframe from DGEobj)
   contrast <- names(DGEobj::getType(dgeObj, type = "topTable"))

   cdfPlot(dgeObj, contrast = contrast[1], title = "My CDF Plot")

   cdfPlot(dgeObj, contrast = contrast[1], title = "My CDF Plot", plotType = "ggplot")
}
}
