% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfPlot.R
\name{cdfPlot}
\alias{cdfPlot}
\title{Create deluxe CDF Plots}
\usage{
cdfPlot(
  contrastDF,
  plotType = "canvasXpress",
  pvalCol = "P.Value",
  pThreshold = 0.01,
  xlab,
  ylab,
  title = NULL,
  insetTitle = NULL,
  symbolSize = c(4, 3),
  symbolShape = c("circle", "circle"),
  symbolColor = c("red3", "deepskyblue4"),
  transparency = 0.7,
  referenceLine = NULL,
  refLineThickness = 3,
  viewportX = 0.15,
  viewportY = 0.85,
  viewportWidth = 0.35,
  pvalMax = 0.1,
  footnote
)
}
\arguments{
\item{contrastDF}{A dataframe with LogRatio and LogIntensity columns and optionally a p-value or FDR column.}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{pvalCol}{Name of the p-value or FDR column (default = "P.Value")}

\item{pThreshold}{Used to color points (default = 0.01)}

\item{xlab}{X axis label (default = "Rank")}

\item{ylab}{Y axis label (default = p-value column name)}

\item{title}{Plot title (Optional)}

\item{insetTitle}{Title for the inset plot (Optional)}

\item{symbolSize}{Size of symbols for Not Significant and  Significant points (default = c(2,1);}

\item{symbolShape}{Shape of the symbols for Not Significant and Significant points (default = c("circle", "circle")
See \url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{Color of symbols for Not Significant and Significant points (default = c("red3", "deepskyblue4"))
See \url{http://research.stowers-institute.org/efg/R/Color/Chart}}

\item{transparency}{Controls the transparency of the plotted points. Value ranges between 0 -1 (default = 0.5)}

\item{referenceLine}{Color for an horizontal line drawn at the p-threshold
(default = NULL; NULL disables, set to desired color to enable)}

\item{refLineThickness}{Set thickness of the reference line (default = 1)}

\item{viewportX}{x-location for the inset plot(default = 0.15)}

\item{viewportY}{y-location for the inset plot(default = 0.85)}

\item{viewportWidth}{width of the inset plot (default = 0.35)}

\item{pvalMax}{Limit the range of the main plot (default = 0.10)}

\item{footnote}{Optional string placed right justified at bottom of plot.}
}
\value{
A list containing main plot, inset plot for both plotType. For plotType ="ggplot" list contains a combined plot which
displays the inset plot in a viewport
}
\description{
CDF plots are a good complement to p-value histograms as a way to evaluate
model performance and examine support for differential expression. Results
are ranked by p-value on the x-axis and the p-value plotted on the y-axis.
Since p-value distributions should be flat, this type of plot should produce a
straight line.  Any observations that fail to meet the null hypothesis will
appear as a break in the line at the low end of the curve.
}
\details{
This function is designed to take topTable dataframes and display the
corresponding CDF plot. Data for the p-values below 0.1 (configurable via
pvalMax argument) are shown in a full size plot. An inset figure shows the
whole p-value scale. Points below 0.01 are a different color by default
(threshold set by pThreshold argument; shape/color attributes customizable
through other arguments).

\strong{Data Structure for the input dataframe:}

The defaults are set for dataframes produced by topTable.  The column
"P.Value" is used by default to accommodate the column names used in topTable
dataframes.  Any other dataframe can be used with by explicitly defining the
p-value column name with the appropriate argument.

Sensible defaults are chosen for symbols (Size, Shape and Color).
There are optional arguments that allow these to be adjusted. A length of 2
is required for these arguments which applies the attributes in
this order: Significant, Not Significant.
}
\examples{
\dontrun{
   # Plot to console (contrastDF is a topTable dataframe)
   cdfPlot(contrastDF, title = "My CDF Plot")
}
}
