% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlot.R
\name{comparePlot}
\alias{comparePlot}
\title{Create formatted scatterplot of a common column from 2 contrasts}
\usage{
comparePlot(
  dgeObj,
  contrasts,
  colorBySigMeasure = TRUE,
  pvalCol = "P.Value",
  valueCol = "logFC",
  plotType = "canvasXpress",
  pThreshold = 0.01,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  referenceLine = "darkgoldenrod1"
)
}
\arguments{
\item{dgeObj}{A DGeobj with atleast 2 contrasts}

\item{contrasts}{The itemNames of a pair of two contrasts in DGEobj that has logFC column.
Optionally add xp and yp columns to hold p-values or FDR values using colorBySigMeasure.}

\item{colorBySigMeasure}{Colors points by significance measures.  (default = TRUE)}

\item{pvalCol}{Name of the p-value column (default = "P.Value")}

\item{valueCol}{Column name of the data for plot (default = "logFC")}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{pThreshold}{Significant value threshold (default = 0.01)}

\item{xlab}{X-axis label (default = first column name)}

\item{ylab}{Y-axis label (default = second column name)}

\item{title}{Plot title (Optional)}

\item{referenceLine}{Color for a slope=1, intercept=0 reference line
(default = "darkgoldenrod1"; NULL disables)}
}
\value{
canvasXpress or ggplot object based on plotType selection
}
\description{
Creates a nicely formatted scatterplot of any common column from any 2 topTable dataframes(defaults to logFC).
 Additionally, if p-values or FDR values and a threshold are supplied,
the plot is color coded to show X unique, Y unique, and
common differentially expressed (DE) genes in different colors.
}
\details{
Other options add an identity line (slope = 1, intercept = 0) and/or a crosshair at (0, 0).

\strong{Data Structure for the input dataframe:}

The x and y values should a common column from two chosen contrasts. By default, the
contrast names will be used for the axis labels. The x and y labels can be changed
using the xlab and ylab arguments.

Optionally, significance measures in the form of p-values or FDR values can be supplied
for X and Y respectively. If provided, these columns \strong{must} be named "xp" and "yp" respectively.
Together with a threshold (which defaults to 0.01), these values will
be used to color code the plot to show X Unique, Y Unique, and Common DE
genes.  Use either p-values or FDR values for the significance columns. Use the
pThreshold argument to set a proper threshold for FDR values.
}
\examples{
\dontrun{
  # Retrieve the first two contrasts from a DGEobj as a list of dataframes (length = 2; named items)
  dgeObj       <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
  contrasts    <- names(DGEobj::getType(dgeObj, "topTable"))[1:2]
  contrastList <- lapply(contrasts, function(x) {
     DGEobj::getItems(dgeObj, x)
   })
   names(contrastList) <- contrasts

  # Draw the plot
  cPlot <- comparePlot(dgeObj, contrasts, title = "Plot Title")
  print(cPlot)

  # Deluxe Plot with bells and whistles.
  myPlot <- comparePlot(dgeObj,
                        contrasts,
                        pThreshold = 0.5,
                        xlab = "x Axis Label",
                        ylab = "y Axis Label",
                        title = "Plot Title",
                        referenceLine = "blue")
}

}
