% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlot.R
\name{comparePlot}
\alias{comparePlot}
\title{Create formatted scatterplot of first two cols of data.frame}
\usage{
comparePlot(
  compareDF,
  plotType = "canvasXpress",
  pThreshold = 0.01,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  symbolSize = c(7, 7, 7, 3),
  symbolShape = c("circle", "circle", "circle", "circle"),
  symbolColor = c("darkgoldenrod1", "deepskyblue4", "red3", "grey25"),
  transparency = 0.5,
  crosshair = "grey50",
  referenceLine = "darkgoldenrod1",
  refLineThickness = 1,
  legendPosition = "right",
  footnote,
  footnoteSize = 3,
  footnoteColor = "black"
)
}
\arguments{
\item{compareDF}{A dataframe with the first two columns representing the x and y variables.
Optionally add xp and yp columns to hold p-values or FDR values.}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{pThreshold}{Used to color points (default = 0.01)}

\item{xlab}{X-axis label (default = first column name)}

\item{ylab}{Y-axis label (default = second column name)}

\item{title}{Plot title (Optional)}

\item{symbolSize}{Size of symbols (default = c(7, 7, 7, 3))}

\item{symbolShape}{Shape of the symbols (default = c("circle", "circle", "circle", "circle")).
For ggplot reference see \url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{c(Common, xUnique, yUnique, NoChange) symbols colors (default = c("darkgoldenrod1", "deepskyblue4", "red3", "grey25"))}

\item{transparency}{Controls the transparency of the plotted points (0-1; default = 0.5)}

\item{crosshair}{Color for the crosshair (default = "grey50">, NULL disables)
See \url{http://research.stowers-institute.org/efg/R/Color/Chart}}

\item{referenceLine}{Color for a slope=1, intercept=0 reference line
(default = "darkgoldenrod1"; NULL disables)}

\item{refLineThickness}{Set thickness for crosshair and referenceLine (default = 1)}

\item{legendPosition}{One of "top", "bottom", "left", "right", "ne", "se", "nw", "sw", NULL.
top/bottom/left/right place the legend outside the figure.  ne/se/nw/sw place the figure
inside the figure. NULL disables the legend (default = "right")}

\item{footnote}{Optional string placed right justified at bottom of plot.}

\item{footnoteSize}{Applies to footnote (default = 3)}

\item{footnoteColor}{Applies to footnote (default = "black")}
}
\value{
canvasXpress or ggplot object based on plotType selection
}
\description{
Creates a nicely formatted scatterplot of the
first two columns in a dataframe.  Several formatting options are
available to enhance the plots.  If p-values or FDR values and a threshold are supplied,
the plot is color coded to show X unique, Y unique, and
common differentially expressed (DE) genes in different colors.
}
\details{
Other options add an identity line (slope = 1, intercept = 0) and/or a crosshair at (0, 0).

\strong{Data Structure for the input dataframe:}

The x and y values should be in the first two columns. By default, their
column names will be used for the axis labels. The x and y labels can be changed
using the xlab and ylab arguments.

Optionally, significance measures in the form of p-values or FDR values can be supplied
for X and Y respectively. If provided, these columns \strong{must} be named "xp" and "yp" respectively.
Together with a threshold (which defaults to 0.01), these values will
be used to color code the plot to show X Unique, Y Unique, and Common DE
genes.  Use either p-values or FDR values for the significance columns. Use the
pThreshold argument to set a proper threshold for FDR values.

Sensible defaults are chosen for symbols (Size, Shape, Color and Fill), but optional
arguments allow much about the symbols to be customized. A length of 4 is
required for these arguments which applies the attributes in this order:
Common, X Unique, Y Unique, and Not Significant.

Note: if p-values or FDR values are not used to color the plot, the X Unique color
values are used.
}
\examples{
\dontrun{
  # Retrieve the first two contrasts from a DGEobj as a list of dataframes (length = 2; named items)
  contrastList <- getType(DGEobj, "topTable")[1:2]

  # Capture the default logFC and P.Value
  compareDat <- comparePrep(contrastList)

  # Switch to an FDR value for the significance measure
  compareDat <- comparePrep(contrastList, significanceCol = "adj.P.Val")

  # Draw the plot
  cPlot <- comparePlot(compareDat, title = "Plot Title")
  print(cPlot)

  # Deluxe Plot with bells and whistles.
  myPlot <- comparePlot(compareDat,
                        pThreshold = 0.5,
                        xlab = "x Axis Label",
                        ylab = "y Axis Label",
                        title = "Plot Title",
                        crosshair = "red",
                        referenceLine = "blue",
                        legendPosition = "right")
}

}
