% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePlot.R
\name{comparePlot}
\alias{comparePlot}
\title{Create formatted scatterplot of first two cols of data.frame}
\usage{
comparePlot(
  compareDF,
  pThreshold = 0.01,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  symbolSize = c(4, 4, 4, 2),
  symbolShape = c(21, 21, 21, 20),
  symbolColor = c("black", "grey0", "grey1", "grey25"),
  symbolFill = c("darkgoldenrod1", "deepskyblue4", "red3", "grey25"),
  alpha = 0.5,
  rugColor = NULL,
  rugAlpha = 1,
  crosshair = "grey50",
  referenceLine = "darkgoldenrod1",
  refLineThickness = 1,
  dens2D = TRUE,
  legendPosition = "right",
  baseFontSize = 12,
  themeStyle = "bw",
  footnote,
  footnoteSize = 3,
  footnoteColor = "black",
  footnoteJust = 1
)
}
\arguments{
\item{compareDF}{A dataframe with the first two columns representing the x and y variables.
Optionally add xp and yp columns to hold p-values or FDR values.}

\item{pThreshold}{Used to color points (Default = 0.01)}

\item{xlab}{X-axis label (Default to first column name)}

\item{ylab}{Y-axis label (Default to second column name)}

\item{title}{Plot title (Optional)}

\item{symbolSize}{Size of symbols; default = c(4, 4, 4, 2)}

\item{symbolShape}{Shape of the symbols; Default = c(21, 21, 21, 20)  (20 = filled circle,
21 = fillable open circle) See \url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{c(Common, xUnique, yUnique, NoChange);
default = c("black", "grey0", "grey1", "grey25")}

\item{symbolFill}{Set the fill color for the symbols. Note only symbols 21-25 are fillable. This will
have no effect on other symbols. Default = c("darkgoldenrod1", "deepskyblue4", "red3", "grey25")}

\item{alpha}{Controls the transparency of the plotted points (0-1; Default = 0.5)}

\item{rugColor}{Specify color for rug density plot along X and Y axes. Set to NULL
to disable rug layer. (Default = NULL)}

\item{rugAlpha}{Sets the transparency for the rug layer. alpha <1.0 takes
a long time to draw, so for a final plot try 0.1 or 0.2 for alpha for more informative rug.}

\item{crosshair}{Color for the crosshair; (Default = "grey50", NULL disables)
See \url{http://research.stowers-institute.org/efg/R/Color/Chart}}

\item{referenceLine}{Color for a slope=1, intercept=0 reference line
(Default = "darkgoldenrod1"; NULL disables)}

\item{refLineThickness}{Set thickness for crosshair and referenceLine (Default = 1)}

\item{dens2D}{Overlay a density2D layer (Default = TRUE) Note: only applies
to simple compare plot without p-value coloring.}

\item{legendPosition}{One of "top", "bottom", "left", "right", "ne", "se", "nw", "sw", NULL.
top/bottom/left/right place the legend outside the figure.  ne/se/nw/sw place the figure
inside the figure. NULL disables the legend. Default = "right"}

\item{baseFontSize}{The smallest size font in the figure in points. Default = 12}

\item{themeStyle}{"bw" or "grey" which corresponds to theme_bw or theme_grey respectively.
Default = "bw"}

\item{footnote}{Optional string placed right justified at bottom of plot.}

\item{footnoteSize}{Applies to footnote. (Default = 3)}

\item{footnoteColor}{Applies to footnote. (Default = "black")}

\item{footnoteJust}{Value 0-1. 0 is left justified, 1 is right justified, 0.5 is centered. (Default = 1)}
}
\value{
ggplot object
}
\description{
Creates a nicely formatted scatterplot of the
first two columns in a dataframe.  Several formatting options are
available to enhance the plots.  If p-values or FDR values and a threshold are supplied,
the plot is color coded to show X unique, Y unique, and
common differentially expressed (DE) genes in different colors.
}
\details{
Other options add an identity line (slope = 1, intercept = 0) and/or a crosshair at (0, 0).

comparePlot has also been implemented with a scalable theme that makes it easy to scale font sizes larger
or smaller for PPT or knitr output.  Add either theme_grey(n) or theme_bw(n),
where n equals the base font size to adjust the text size (12 works well for knitr, 18 or 24 works well
for PPT), e.g. MyPlot + theme_grey(18).  However, theme_grey() and theme_bw() have the
side effect of resetting the legend position, and should not be used if the custom legend locations
need to be preserved. Instead, utilize the baseFontSize
argument in this function to set font sizes and preserve a custom legend location.

\strong{Data Structure for the input dataframe:}

The x and y values should be in the first two columns. By default, their
column names will be used for the axis labels. The x and y labels can be changed
using the xlab and ylab arguments.

Optionally, significance measures in the form of p-values or FDR values can be supplied
for X and Y respectively. If provided, these columns \strong{must} be named "xp" and "yp" respectively.
Together with a threshold (which defaults to 0.01), these values will
be used to color code the plot to show X Unique, Y Unique, and Common DE
genes.  Use either p-values or FDR values for the significance columns. Use the
pThreshold argument to set a proper threshold for FDR values.

Sensible defaults are chosen for symbols (Size, Shape, Color and Fill), but optional
arguments allow much about the symbols to be customized. A length of 4 is
required for these arguments which applies the attributes in this order:
Common, X Unique, Y Unique, and Not Significant.

Note: if p-values or FDR values are not used to color the plot, the X Unique color
values are used.
}
\examples{
\dontrun{
  # Retrieve the first two contrasts from a DGEobj as a list of dataframes (length = 2; named items)
  contrastList <- getType(DGEobj, "topTable")[1:2]

  # Capture the default logFC and P.Value
  compareDat <- comparePrep(contrastList)

  # Switch to an FDR value for the significance measure
  compareDat <- comparePrep(contrastList, significanceCol = "adj.P.Val")

  # Draw the plot
  cPlot <- comparePlot(compareDat, title = "Plot Title")
  print(cPlot)

  # Deluxe Plot with bells and whistles.
  myPlot <- comparePlot(compareDF,
                        pThreshold = 0.5,
                        xlab = "x Axis Label",
                        ylab = "y Axis Label"
                        title = "Plot Title",
                        crosshair = "red",
                        referenceLine = "blue",
                        legendPosition = "nw")
}

}
