% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{Create limma MDS plot}
\usage{
ggplotMDS(
  dgeObj,
  plotType = "canvasXpress",
  designTable = "design",
  colorBy = "ReplicateGroup",
  shapeBy = NULL,
  sizeBy = NULL,
  top = Inf,
  labels = "ReplicateGroup",
  title,
  xlab,
  ylab,
  hlineIntercept,
  vlineIntercept
)
}
\arguments{
\item{dgeObj}{A DGEobj object with a DGEList and design table (required)}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress)}

\item{designTable}{Name of the design table object (default = design)}

\item{colorBy}{A column name in the design table. Points are colored by the values in that column (default = ReplicateGroup)}

\item{shapeBy}{A column name in the design table. Points are shaped by the values in that column (default = NULL)}

\item{sizeBy}{A column name in the design table. Points are sized by the values in that column (default = NULL)}

\item{top}{Number of most variant genes to include (default = Inf)}

\item{labels}{A column name in the design table. Text labels for the samples. These should be short
abbreviations of the sample identifiers.
Default = ReplicateGroup or rownames of dgeObj. Set to NULL to disable
text labels.}

\item{title}{A title for the plot. (Optional)}

\item{hlineIntercept}{Y intercept of horizontal line (Optional)}

\item{vlineIntercept}{X intercept of vertical line (Optional)}
}
\value{
A list with two elements, the ggplot object and the MDS object returned
   by the plotMDS() function.
}
\description{
This is a wrapper around the plotMDS function that generates the plot with
canvasXpress or ggplot2 instead of base graphics.
}
\details{
colorBy, shapeBy, and sizeBy are grouping variables that encode group info by
color, shape, or size.  These are input as column names in the designTable.

The underlying limma::plotMDS() function uses a default of top = 500 to use the top 500
highest fold change genes for the analysis. Based on observed speed tests,
top = Inf has been utilized as the default for this function, as it was shown to quickly produce
a more stable result. However, this is configurable using the top argument, which
allows for selection of a number close the number of differential genes in the
input data.
}
\examples{
\dontrun{
     # Plot the first two dimensions using all genes
     myMDS_cxplot <- ggplotMDS(MyDGEobj)

     # MDSplot - ggplot
     myMDS_ggplot <- ggplotMDS(MyDGEobj, plotType = "ggplot")
     myMDS_ggplot[[1]]
}

}
