% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{Create limma MDS plot}
\usage{
ggplotMDS(
  DGEdata,
  plotType = "canvasXpress",
  designTable = "design",
  colorBy = "ReplicateGroup",
  shapeBy = NULL,
  sizeBy = NULL,
  top = Inf,
  labels = "ReplicateGroup",
  labelSize = 3,
  title,
  hlineIntercept,
  vlineIntercept,
  reflineColor = "red",
  reflineSize = 0.5,
  symShape = "circle",
  symSize = 10,
  transparency = 0.7,
  dim.plot = c(1, 2)
)
}
\arguments{
\item{DGEdata}{A DGEobj that contains a DGEList OR a log2cpm matrix. (Required)}

\item{plotType}{Plot type must be canvasXpress or ggplot (Default to canvasXpress).}

\item{designTable}{Name of the design table object}

\item{colorBy}{A column name in the design table.Points are colored by the values in that column (Required)}

\item{shapeBy}{A column name in the design table.Points are shaped by the values in that column (Optional)}

\item{sizeBy}{A column name in the design table.Points are sized by the values in that column (Optional)}

\item{top}{Number of most variant genes to include (Default = Inf)}

\item{labels}{A column name in the design table. Text labels for the samples. These should be short
abbreviations of the sample identifiers.
Default = ReplicateGroup or rownames of DGEdata. Set to NULL to disable
text labels.}

\item{labelSize}{Control size for the text labels in the plot,}

\item{title}{A title for the plot. (Optional)}

\item{hlineIntercept}{Y intercept of horizontal line (Optional)}

\item{vlineIntercept}{X intercept of vertical line (Optional)}

\item{reflineColor}{Color for the horizontal and vertical reference lines
(default = "darkgoldenrod1")}

\item{reflineSize}{Thickness of the reference lines (default = 0.5)}

\item{symShape}{Set the default shape of the symbols if not mapped to a column (default = circle)}

\item{symSize}{Set the default size of the symbols if not mapped to a column
(default = 10)}

\item{transparency}{Set transparency (default = 0.7)}

\item{dim.plot}{Define which dimension to plot. dim.plot should a numeric vector of
length 2 and should be lesser than the number of columns in DGEobj. (default = c(1,2))}
}
\value{
A list with two elements, the ggplot object and the MDS object returned
   by the plotMDS() function.
}
\description{
This is a wrapper around the plotMDS function that generates the plot with
canvasXpress or ggplot2 instead of base graphics.
}
\details{
colorBy, shapeBy, and sizeBy are grouping variables that encode group info by
color, shape, or size.  These are input as column names in the designTable.

The underlying limma::plotMDS() function uses a default of top = 500 to use the top 500
highest fold change genes for the analysis. Based on observed speed tests,
top = Inf has been utilized as the default for this function, as it was shown to quickly produce
a more stable result. However, this is configurable using the top argument, which
allows for selection of a number close the number of differential genes in the
input data.
}
\examples{
\dontrun{
     # Plot the first two dimensions using all genes
     myMDS_cxplot <- ggplotMDS(MyDGEobj)

     # Plot the 2nd and 3rd dimensions using the top 1000 genes
     myMDS_cxplot <- ggplotMDS(MyDGEobj, dim.plot = c(2, 3))
     myMDS_cxplot[[1]]

     # MDSplot - ggplot
     myMDS_ggplot <- ggplotMDS(MyDGEobj, plotType = "ggplot")
     myMDS_ggplot <- ggplotMDS(MyDGEobj, plotType = "ggplot", dim.plot = c(2, 3))
     myMDS_ggplot[[1]]
}

}
