% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{Create limma MDS plot using ggplot2}
\usage{
ggplotMDS(
  DGEdata,
  colorBy,
  shapeBy,
  sizeBy,
  top = Inf,
  labels,
  labelSize,
  title,
  textColor = "blue2",
  hlineIntercept,
  vlineIntercept,
  reflineColor = "darkgoldenrod1",
  reflineSize = 0.5,
  baseFontSize = 12,
  themeStyle = "grey",
  symShape = 16,
  symSize = 5,
  symFill = "blue2",
  symColor = "blue2",
  alpha = 0.7,
  shapes,
  colors,
  dim.plot = c(1, 2),
  colorName,
  shapeName,
  sizeName
)
}
\arguments{
\item{DGEdata}{A DGEList object taken after normalization
OR a DGEobj that contains a DGEList OR a log2cpm matrix. (Required)}

\item{colorBy}{A grouping vector to color by (e.g. ReplicateGroup) (Required)}

\item{shapeBy}{A grouping vector to map to shape (Optional)}

\item{sizeBy}{A numeric vector to define point size (Optional)}

\item{top}{Number of most variant genes to include (Default = Inf)}

\item{labels}{Text labels for the samples. These should be short
abbreviations of the sample identifiers.
Default = ReplicateGroup or rownames of DGEdata. Set to NULL to disable
text labels.}

\item{labelSize}{Control size for the text labels in the plot,}

\item{title}{A title for the plot. (Optional)}

\item{textColor}{Color for the text labels in the plot (Default = "blue2")}

\item{hlineIntercept}{Y intercept of horizontal line (Optional)}

\item{vlineIntercept}{X intercept of vertical line (Optional)}

\item{reflineColor}{Color for the horizontal and vertical reference lines
(Default = "darkgoldenrod1")}

\item{reflineSize}{Thickness of the reference lines (Default = 0.5)}

\item{baseFontSize}{Base font size for the plot (Default = 12)}

\item{themeStyle}{One of "grey" or "bw" (Default = "grey")}

\item{symShape}{Set the default shape of the symbols if not mapped to a column (Default = 19, solid circle)}

\item{symSize}{Set the default size of the symbols if not mapped to a column
(Default = 5)}

\item{symFill}{Set color for the fill on open symbols (Default = "blue2")}

\item{symColor}{Set color for solid symbols or outline for open symbols
(Default = "blue2")}

\item{alpha}{Set transparency (Default = 0.7)}

\item{shapes}{A vector of shapes to override the default 8 shapes used in shapeBy (optional)}

\item{colors}{A color pallet to substitute for the default 8 color pallet used by colorBy (optional)}

\item{dim.plot}{Define which dimension to plot (Default = c(1,2))}

\item{colorName}{Legend title for color (optional)}

\item{shapeName}{Legend title for shape (optional)}

\item{sizeName}{Legend title for size (optional)}
}
\value{
A list with two elements, the ggplot object and the MDS object returned
   by the plotMDS() function.
}
\description{
This is a wrapper around the plotMDS function that generates the plot with
ggplot2 instead of base graphics.
}
\details{
colorBy, shapeBy, and sizeBy are grouping variables that encode group info by
color, shape, or size.  These are vectors that must be the same length as
ncol(DGEdata). colorBy and sizeBy will plot as continuous color or size
changes if a numeric vector is used. Convert the vector to a factor to
treat as groups instead of continuous.

The underlying limma::plotMDS() function uses a default of top = 500 to use the top 500
highest fold change genes for the analysis. Based on observed speed tests,
top = Inf has been utilized as the default for this function, as it was shown to quickly produce
a more stable result. However, this is configurable using the top argument, which
allows for selection of a number close the number of differential genes in the
input data.
}
\examples{
\dontrun{
     # Plot the first two dimensions using all genes
     myMDS <- ggplotMDS(MyDGEList)

     # Plot the 2nd and 3rd dimensions using the top 1000 genes
     myMDS <- ggplotMDS(myDGEList, dim.plot = c(2, 3) ndim = 3)
     myMDS[[1]]
}

}
