% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioPlot.R
\name{logRatioPlot}
\alias{logRatioPlot}
\title{Plot logRatio contrasts}
\usage{
logRatioPlot(
  contrastsDF,
  plotType = "canvasXpress",
  facetColname,
  xColname,
  yColname = "logFC",
  CI.R_colname = "CI.R",
  CI.L_colname = "CI.L",
  plotCategory = "bar",
  refLine = TRUE,
  refLineColor = "red",
  xlab = xColname,
  ylab = yColname,
  title = NULL,
  barColor = "dodgerblue4",
  barTransparency = 1,
  barWidth = 0.9,
  pointColor = "dodgerblue4",
  pointShape = "circle",
  pointTransparency = 1,
  pointSize = 2,
  lineLayer = FALSE,
  lineColor = "dodgerblue4",
  lineSize = 1,
  lineFit = "loess",
  facet = TRUE,
  facetCol,
  labelAngle = 45,
  axisFree = TRUE
)
}
\arguments{
\item{contrastsDF}{A tidy dataframe of data to plot (Required) (see ?tidyContrasts).}

\item{facetColname}{Define the column name to separate plots (Required) (e.g. GeneID).}

\item{xColname}{Define the column name to group boxplots by (Required) (e.g. Contrast).}

\item{yColname}{Define the column name for the output of the boxplots (default = "logFC")}

\item{CI.R_colname}{Define name of the CI high value (default = "CI.R")}

\item{CI.L_colname}{Define name of the CI low value (default =  "CI.L")}

\item{plotCategory}{One of "bar" or "point" (default = "bar")}

\item{refLine}{Adds a horizontal line at y = 0 (default = TRUE)}

\item{refLineColor}{Color for the reference line (default = "red")}

\item{xlab}{X axis label (defaults to xColname)}

\item{ylab}{Y axis label (defaults to yColname)}

\item{title}{Plot title (Optional)}

\item{barColor}{Color for the bar outline (default = "dodgerblue4")}

\item{barTransparency}{Transparency for the box layer (default = 1)}

\item{barWidth}{Set the bar width (default = 0.8)}

\item{pointColor}{Color for the point layer (default = "grey30")}

\item{pointShape}{Shape for the point layer (default = "circle")}

\item{pointSize}{Size of the points (default = 4)}

\item{lineLayer}{Add a fitted line layer (default = FALSE)}

\item{lineColor}{Color of the line fit (default = "dodgerblue4")}

\item{lineSize}{Size of the line fit (default = 1)}

\item{lineFit}{Type of fit to use.  One of c("auto", "lm", "glm", "gam",
"loess"). (default = "loess")}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE)}

\item{facetCol}{Explicitly set the number of rows for the facet plot. default
behavior will automatically set the columns. (default = ceiling(sqrt(length(unique(contrastsDF[facetCol])))))}

\item{labelAngle}{Angle to set the sample labels on the X axis (Default =  45; Range = 0-90)}

\item{axisFree}{Specify same scale or independent scales for each subplot (Default = TRUE;
Allowed values: TRUE and FALSE)}
}
\value{
ggplot object. If facet = TRUE (default), returns a faceted ggplot object. If
  facet = FALSE, returns a list of ggplot objects indexed
  by observation (gene) names.
}
\description{
Intended to plot a set of contrast results, one plot for each gene of
interest. Input is a tidy datafile constructed from topTable output and
requires logFC, CI.L, and CI.R columns as well as a gene identifier of choice.
Outputs a ggplot2 object faceted by the facetColname or a list of individual
ggplots, one for each facetColname value (typically gene).
}
\examples{
\dontrun{
  # DGEobj example
  # Put contrasts in tidy format keeping logFC, and confidence limits contrastsDF
  tidyDat <- tidyContrasts(DGEobj,
                           rownameColumn = "EnsgID",
                           includeColumns = c("logFC", "CI.R", "CI.L"))

  # Add gene symbols from geneData
  ens2genesym <- DGEobj$geneData \%>\%
                 rownames_to_column(var = "EnsgID") \%>\%
                 select(EnsgID, GeneSymbol = GeneName)
  tidyDat <- left_join(tidyDat, ens2genesym)

  # Filter for a small set of genes of interest
  idx <- stringr::str_detect(tidyDat$GeneSymbol, "^PPAR")
  tidyDat <- tidyDat[idx,]

  # Simple barplot
  logRatioPlot(tidyDat,
               facetColname = "GeneSymbol",
               xColname = "Contrast",
               facetCol = 2)

  # Lineplot with some options
  logRatioPlot(tidyDat,
               plotCategory = "point",
               facetColname = "GeneSymbol",
               xColname = "Contrast",
               facetCol = 4,
               axisFree = FALSE,
               facet = TRUE,
               title = "Test",
               pointSize = 4,
               lineLayer = TRUE,
               lineSize = 0.1,
               labelAngle = 60)
}

}
