% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioPlot.R
\name{logRatioPlot}
\alias{logRatioPlot}
\title{Plot logRatio contrasts}
\usage{
logRatioPlot(
  dgeObj,
  plotType = "canvasXpress",
  facetColname = "rgd_symbol",
  xColname = "Contrast",
  yColname = "logFC",
  CI.R_colname = "CI.R",
  CI.L_colname = "CI.L",
  plotCategory = "bar",
  refLine = TRUE,
  refLineColor = "red",
  xlab = xColname,
  ylab = yColname,
  title = NULL,
  barColor = "dodgerblue4",
  barTransparency = 1,
  pointColor = "dodgerblue4",
  pointShape = "circle",
  pointTransparency = 1,
  pointSize = 2,
  facet = TRUE,
  labelAngle = 45,
  axisFree = TRUE
)
}
\arguments{
\item{dgeObj}{A DGEobj that contains a contrast or more dataframe.}

\item{plotType}{Plot type must be canvasXpress or ggplot (Default to canvasXpress).}

\item{facetColname}{Define the column name to separate plots from geneData (default = "rgd_symbol").}

\item{xColname}{Define name for the computed column to group boxplots by (Required) (default = "Contrast").}

\item{yColname}{Define the column name for the output of the boxplots (default = "logFC")}

\item{CI.R_colname}{Define name of the CI high value (default = "CI.R")}

\item{CI.L_colname}{Define name of the CI low value (default =  "CI.L")}

\item{plotCategory}{One of "bar" or "point" (default = "bar")}

\item{refLine}{Adds a horizontal line at y = 0 (default = TRUE)}

\item{refLineColor}{Color for the reference line (default = "red")}

\item{xlab}{X axis label (defaults to xColname)}

\item{ylab}{Y axis label (defaults to yColname)}

\item{title}{Plot title, set to NULL to disable (Optional, default is NULL)}

\item{barColor}{Color for the bar outline (default = "dodgerblue4")}

\item{barTransparency}{Transparency for the bar layer (default = 1)}

\item{pointColor}{Color for the point layer (default = "grey30")}

\item{pointShape}{Shape for the point layer, for ggplot plotType (default = "circle")}

\item{pointTransparency}{Transparency for the box layer (default = 1)}

\item{pointSize}{Size of the points, for ggplot plotType (default = 4)}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE)}

\item{labelAngle}{Angle to set the sample labels on the X axis (Default =  45; Range = 0-90)}

\item{axisFree}{Specify same scale or independent scales for each subplot (Default = TRUE;
Allowed values: TRUE and FALSE)}
}
\value{
canvasXpress (the default) or a ggplot object:
#' \itemize{
  \item If \code{facet = TRUE} \emph{(default)}, returns a faceted canvasXpress or ggplot object.
  \item If \code{facet = FALSE}, returns a list of canvasXpress or ggplot objects indexed by observation (gene) names.
}
}
\description{
Intended to plot a set of contrast results, one plot for each gene of
interest.
}
\details{
Input is a DGEobj object that must contain  topTable object with at least one contrast dataframe.
Also, it requires \strong{logFC}, \strong{CI.L},
and \strong{CI.R} columns as well as a gene identifier of choice.

Outputs a canvasXpress \emph{(the default)} or ggplot2 object faceted by the facetColname when
 \code{facet} parameter is set to \strong{TRUE} \emph{(the default)}.

If \code{facet} parameter is set to \strong{FALSE} the output will be a list of individual canvasXpress or
ggplots, one for each facetColname value \emph{(typically gene)}.
}
\examples{
\dontrun{
  # DGEobj example, subset DGEobj
  t_obj1_subset <- subset(t_obj1, row = c(1:6))


  # Simple barplot
 logRatioPlot(t_obj1_subset,
              facetColname = "rgd_symbol",
              xColname = "Contrast")

  # Lineplot with some options
 logRatioPlot(t_obj1_subset,
              plotCategory = "point",
              facetColname = "rgd_symbol",
              xColname = "Contrast",
              axisFree = FALSE,
              facet = TRUE,
              title = "Test",
              pointSize = 4,
              labelAngle = 60)
}

}
