% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRatioPlot.R
\name{logRatioPlot}
\alias{logRatioPlot}
\title{Plot logRatio contrasts}
\usage{
logRatioPlot(
  dgeObj,
  plotType = "canvasXpress",
  facetColname = "rgd_symbol",
  xColname = "Contrast",
  yColname = "logFC",
  CI.R_colname = "CI.R",
  CI.L_colname = "CI.L",
  plotCategory = "bar",
  refLine = TRUE,
  xlab = xColname,
  ylab = yColname,
  title = NULL,
  facet = TRUE,
  labelAngle = 45,
  axisFree = TRUE
)
}
\arguments{
\item{dgeObj}{A DGEobj that contains topTable objects}

\item{plotType}{Plot type must be canvasXpress or ggplot (default to canvasXpress).}

\item{facetColname}{Column name to separate plots from geneData (default = "rgd_symbol").}

\item{xColname}{The computed column to group boxplots by (default = "Contrast").}

\item{yColname}{Column name for the output of the boxplots (default = "logFC")}

\item{CI.R_colname}{Name of the CI high value (default = "CI.R")}

\item{CI.L_colname}{Name of the CI low value (default =  "CI.L")}

\item{plotCategory}{One of "bar" or "point" (default = "bar")}

\item{refLine}{Adds a horizontal line at y = 0 (default = TRUE)}

\item{xlab}{X axis label (defaults to xColname)}

\item{ylab}{Y axis label (defaults to yColname)}

\item{title}{Plot title, set to NULL to disable (Optional, default = NULL)}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE)}

\item{labelAngle}{Angle (0-90) to set the sample labels on the X axis (default =  45)}

\item{axisFree}{Specify same scale or independent scales for each subplot (default = TRUE)}
}
\value{
canvasXpress (the default) or a ggplot object:
#' \itemize{
  \item If \code{facet = TRUE} \emph{(default)}, returns a faceted canvasXpress or ggplot object.
  \item If \code{facet = FALSE}, returns a list of canvasXpress or ggplot objects indexed by observation (gene) names.
}
}
\description{
Intended to plot a set of contrast results, one plot for each gene of
interest.
}
\details{
Input is a DGEobj object that must contain  topTable object with at least one contrast dataframe.
Also, it requires \strong{logFC}, \strong{CI.L},
and \strong{CI.R} columns as well as a gene identifier of choice.

Outputs a canvasXpress \emph{(the default)} or ggplot2 object faceted by the facetColname when
 \code{facet} parameter is set to \strong{TRUE} \emph{(the default)}.

If \code{facet} parameter is set to \strong{FALSE} the output will be a list of individual canvasXpress or
ggplots, one for each facetColname value \emph{(typically gene)}.
}
\examples{
\dontrun{
  # DGEobj example, subset DGEobj
  t_obj1 <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
  t_obj1_subset <- subset(t_obj1, row = c(1:6))

  # Simple barplot
 logRatioPlot(t_obj1_subset,
              facetColname = "rgd_symbol",
              xColname = "Contrast")

  # Point plot with some options
 logRatioPlot(t_obj1_subset,
              plotCategory = "point",
              facetColname = "rgd_symbol",
              xColname = "Contrast",
              axisFree = FALSE,
              facet = TRUE,
              title = "Test",
              labelAngle = 60)
}

}
