% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDGEobj.R
\name{mapDGEobj}
\alias{mapDGEobj}
\title{Plot a network of DGEobj relationships}
\usage{
mapDGEobj(dgeObj, plotType = "canvasXpress", directed = TRUE)
}
\arguments{
\item{dgeObj}{DGEobj to find the parent/child relationships between data items.}

\item{plotType}{Plot type must be canvasXpress or igraph (default = canvasXpress).}

\item{directed}{Only for igraph. Indicates if the graph should
be directed or not. (default = TRUE)}
}
\value{
A class igraph network object or a canvasXpress network plot.
}
\description{
Reads a DGEobj and produces a node pair file or a graph defining parent/child relationships between
data items in the DGEobj.
}
\examples{
\dontrun{
  # Prepare canvasxpress network plot
  mynet <- mapDGEobj(dgeObj)

  # Prepare an iGraph object for plotting
  mynet <- mapDGEobj(dgeObj, plotType = "ggplot")
  plot(mynet)
  }

}
