% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDGEobj.R
\name{mapDGEobj}
\alias{mapDGEobj}
\title{Construct DGEobj relationships}
\usage{
mapDGEobj(dgeObj, plotType = "canvasXpress", directed = TRUE)
}
\arguments{
\item{dgeObj}{DGEobj to find the parent/child relationships between data items.}

\item{plotType}{Must be canvasXpress or ggplot (default = canvasXpress).}

\item{directed}{Only applies to ggplot. Indicates if the graph should
be directed or not. (default = TRUE)}
}
\value{
A class ggplot object or a canvasXpress network plot.
}
\description{
Reads a DGEobj and visualizes parent/child relationships between
data items in a DGEobj.
}
\examples{
\dontrun{
  # canvasXpress network plot
  mapDGEobj(dgeObj)

  # ggplot object
  mapDGEobj(dgeObj, plotType = "ggplot")
  }

}
