% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDGEobj.R
\name{mapDGEobj}
\alias{mapDGEobj}
\title{Plot network of DGEobj relationships}
\usage{
mapDGEobj(dgeObj, directed = TRUE)
}
\arguments{
\item{dgeObj}{DGEobj to find the parent/child relationships between data items.}

\item{directed}{Passed to igraph::graph_from_data_frame. Indicates if the graph should
be directed or not. Default = TRUE.}
}
\value{
A class igraph network object.
}
\description{
Reads a DGEobj and produces a node pair file defining parent/child relationships between
data items in the DGEobj.
}
\examples{
\dontrun{
  library(igraph)
  library(RColorBrewer)

  # Prepare an iGraph object for plotting
  mynet <- mapDGEobj(dgeObj)

  # Define a layout
  lay.sug <- layout_with_sugiyama(mynet)

  # 2D Plot
  plot(mynet,
       edge.arrow.size = .3,
       vertex.color = "dodgerblue2",
       vertex.frame.color = "dodgerblue4",
       layout = lay.sug$layout)

  # 2D Plot; colorby the basetype attribute
  pal <- brewer.pal(length(unique(V(mynet)$basetype)), "Set1")
  myPallet <- pal[as.numeric(as.factor(vertex_attr(mynet, "basetype")))]

  plot(mynet,
       edge.arrow.size = .3,
       vertex.color = myPallet,
       vertex.label.family = "Helvetica",
       layout=lay.sug$layout)

  # 2D Interactive plot
  plotHandle <- tkplot(net,
                       vertex.color = myPallet,
                       vertex.frame.color = "dodgerblue4",
                       canvas.width = 800,
                       canvas.height = 800,
                       layout = lay.sug$layout)
  tk_close(plotHandle)
}

}
