% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot.R
\name{obsPlot}
\alias{obsPlot}
\title{obsPlot: Plots a summary (box/violin) plot the for given gene}
\usage{
obsPlot(
  dgeObj,
  plotType = "canvasXpress",
  countsMatrix = "counts",
  convertCounts = NULL,
  convert_geneLength,
  convert_log = FALSE,
  convert_normalize = "none",
  convert_prior.count = NULL,
  designTable = "design",
  group = "ReplicateGroup",
  violinLayer = FALSE,
  showPoints = TRUE,
  xlab,
  ylab,
  title,
  facet = TRUE,
  axisFree = TRUE
)
}
\arguments{
\item{dgeObj}{A DGEObject. The countsMatrix in the DGEObject is extracted to plot the data. (required)}

\item{plotType}{Can be canvasXpress or ggplot (default = canvasXpress)}

\item{countsMatrix}{Name of the counts matrix in the DGEObj which will be used to render the plot.(default = counts)}

\item{convertCounts}{A flag to indicate if counts matrix need to be converted or taken as is. Default value is NULL. This indicates
countsMatrix need to be taken as is. To convert the counts matrix, specify the desired unit. Supported units include CPM,FPKM, FPK and TPM.
This parameter is passed to DGEobj.utils::convertCounts (default = NULL)}

\item{convert_geneLength}{Parameter to pass to DGEobj.utils::convertCounts. (optional).}

\item{convert_log}{Parameter to pass to DGEobj.utils::convertCounts. (default = FALSE)}

\item{convert_normalize}{Parameter to pass to DGEobj.utils::convertCounts. (default = none)}

\item{convert_prior.count}{Parameter to pass to DGEobj.utils::convertCounts. (default = NULL)}

\item{designTable}{Name of the design table in the DGEObj from which the grouping column will be extracted. (default = design)}

\item{group}{Define the column name to group boxplots by (typically a replicate group column) (required)}

\item{violinLayer}{Adds a violin layer (default = FALSE)}

\item{showPoints}{Shows the data points on the plot (default = TRUE)}

\item{xlab}{X axis label (defaults to group column name if not specified)}

\item{ylab}{Y axis label (defaults to value column name if not specified)}

\item{title}{Plot title (optional)}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE). It is recommended to facet no more than 40 plots. If 40 of more plots are needed, please set this argument to FALSE}

\item{axisFree}{Specify same scale or independent scales for each subplot (default = TRUE;
Allowed values: TRUE or FALSE)}
}
\value{
Plot of type canvasXpress or ggplot. If Facet = TRUE (default) returns a faceted object. If
  facet = FALSE, returns a list of objects indexed
  by observation (gene) names.
}
\description{
Provides a summary plot (box/violin) for each observation (gene), showing data for each
experiment group.
}
\details{
The plot can optionally include boxplot, violinplot or both and can additionally choose to display points on the plot.

By default, the violinLayer is not displayed and only the boxplot, points and the mean on the plots can be seen.
Also, by default, the plots are faceted.

Faceting the plot can be turned off to return a list of individual plots for each gene. Input is a DGEobj with a
Counts Matrix. User can also provide input parameters to convert the counts matrix to other desired units.
}
\examples{
\dontrun{
  #subset the DGEobject to only 6 genes before plotting
  DGEobj6 <- DGEobj[1:6,]

  # Faceted boxplot
  obsPlot(DGEobj6,
          designTable = "design",
          group = "replicategroup",
          countsMatrix = "counts")

  # Faceted violin plot
  obsPlot(DGEobj6,
           violinLayer = TRUE,
           designTable = "design",
           group = "replicategroup",
           countsMatrix = "counts")

  # Return a list of plot for each individual gene
  myplots <- obsPlot(DGEobj6,
                     designTable = "design",
                     group = "replicategroup",
                     countsMatrix = "counts",
                     facet = FALSE)
  # Plot one from the list
  myplots[[2]]

  #ggplot
  obsPlot(DGEobj6,
          designTable = "design",
          group = "replicategroup",
          countsMatrix = "counts",
          plotType = "ggplot")
}

}
\seealso{
\link[DGEobj.utils]{convertCounts}
}
