% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot.R
\name{obsPlot}
\alias{obsPlot}
\title{Function obsPlot}
\usage{
obsPlot(
  data,
  plotType = "canvasXpress",
  plotByCol,
  groupCol,
  valueCol,
  groupOrder = unique(as.character(data[, groupCol, drop = TRUE])),
  boxLayer = TRUE,
  violinLayer = FALSE,
  pointLayer = TRUE,
  meanLayer = TRUE,
  xlab = groupCol,
  ylab = valueCol,
  title,
  boxColor = "deepskyblue3",
  boxTransparency = 0.5,
  boxNotch = FALSE,
  boxNotchWidth = 0.8,
  violinColor = "goldenrod1",
  violinTransparency = 0.5,
  pointColor = "dodgerblue4",
  pointShape = 21,
  pointTransparency = 1,
  pointSize = 2,
  pointJitter = 0,
  meanColor = "goldenrod1",
  meanOutlineColor = "red2",
  meanShape = 22,
  meanTransparency = 0.7,
  meanSize = 3,
  facet = TRUE,
  facetRow,
  xAngle = 30,
  scales = "free_y"
)
}
\arguments{
\item{data}{A tidy dataframe of intensity data with row and colnames (required)}

\item{plotByCol}{Define the column name to separate plots (typically a gene ID column) (required)}

\item{groupCol}{Define the column name to group boxplots by (typically a replicate group column) (required)}

\item{valueCol}{Define the column of values for plotting (typically a log intensity measure) (required)}

\item{groupOrder}{Define the order for the groups in each plot.  Should
contain values in unique(data$group) listed in the order that you want the
groups to appear in the plot. (optional; default = unique(data[groupCol]))}

\item{boxLayer}{Adds a boxplot layer (Default = TRUE)}

\item{violinLayer}{Adds a violin layer (Default = FALSE)}

\item{pointLayer}{Adds a point layer (Default = TRUE)}

\item{meanLayer}{Adds a mean layer (Default = TRUE)}

\item{xlab}{X axis label (defaults to groupCol)}

\item{ylab}{Y axis label (defaults to valueCol)}

\item{title}{Plot title (optional)}

\item{boxColor}{Color for the boxplot layer (Default = "deepskyblue3")}

\item{boxTransparency}{Transparency for the box layer (Default = 0.5)}

\item{boxNotch}{Turn on/off confidence interval notches on boxplots (Default = FALSE)}

\item{boxNotchWidth}{Set the width of box notches (0-1) (Default = 0.8)}

\item{violinColor}{Color for the violin layer (Default = "yellow")}

\item{violinTransparency}{Transparency for the violin layer (Default = 0.5)}

\item{pointColor}{Color for the point layer (Default = "dodgerblue4")}

\item{pointShape}{Shape for the point layer (Default = 21; fillable circle)}

\item{pointTransparency}{Transparency for the box layer (Default = 1)}

\item{pointSize}{Size of the points (Default = 4)}

\item{pointJitter}{Amount to jitter the points (Default = 0) Try 0.2 if you
have a lot of points.}

\item{meanColor}{Color for the mean layer (Default = "goldenrod1")}

\item{meanOutlineColor}{Fill color for the mean layer (Default = "red2")}

\item{meanShape}{Shape for the mean layer (Default = 21; fillable circle)}

\item{meanTransparency}{Transparency for the mean layer (Default = 0.7)}

\item{meanSize}{Size of the mean points (Default = 3)}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (Default = TRUE)}

\item{facetRow}{Explicitly set the number of rows for the facet plot.
Default behavior will automatically set the columns. (Default = NULL)}

\item{xAngle}{Angle to set the sample labels on the X axis. (Default =  30; Range = 0-90)}

\item{scales}{Specify same scales or independent scales for each subplot (Default = "free_y";
Allowed values: "fixed", "free_x", "free_y", "free")}

\item{boxFill}{Fill Color for the boxplot layer (Default = "deepskyblue3")}

\item{violinFill}{Fill Color for the violin (Default = "yellow")}

\item{pointFill}{Fill color for the point layer (Default = "dodgerblue4")}
}
\value{
Plot of type canvasXpress or ggplot. If Facet = TRUE (default) returns a faceted object. If
  facet = FALSE, returns a list of objects indexed
  by observation (gene) names.
}
\description{
Provides a summary plot (box/violin/points) for each observation (gene), showing data for each
experiment group. The plot can optionally include one or more of the
following layers: boxplot, violin plot, individual points, and/or mean of all
points.  The layers are built up in the order listed with user settable
transparency, colors etc.  By default, the boxplot, point, and mean layers are
active. Also, by default, the plots are faceted.  Faceting the plot can be turned
off to return a list of individual ggplot graphics for each gene.
}
\details{
Input is a tidy dataframe of intensity data.
}
\examples{
\dontrun{
  # Simulate some data with row and colnames
  groups <- paste("group", factor(rep(1:4, each = 100)), sep = "")
  x <- matrix(rnorm(2400, mean = 10), ncol = length(groups))
  colnames(x) <- paste("sample", 1:ncol(x), sep = "")
  rownames(x) <- paste("gene", 1:nrow(x), sep="")
  # Reformat into tidy dataframe
  tidyInt <- tidyIntensity(x)

 # Or get data from a DGEobj with RNA-Seq data
 log2cpm <- convertCounts(dgeObj$counts, unit = "cpm", log = TRUE, normalize = "tmm")
 tidyInt <- tidyIntensity(log2cpm,
                          rowidColname = "GeneID",
                          keyColname = "Sample",
                          valueColname = "Log2CPM",
                          group = dgeObj$design$ReplicateGroup)

  # Faceted boxplot
  obsPlot(tidyInt,
           plotByCol = "GeneID",
           groupCol = "group",
           valueCol ="Log2CPM",
           pointJitter = 0.1,
           facetRow = 2)

  # Faceted violin plot
  obsPlot(tidyInt,
           plotByCol = "GeneID",
           violinLayer = TRUE,
           boxLayer = FALSE,
           groupCol="group",
           valueCol = "Log2CPM",
           pointJitter = 0.1,
           facetRow = 2)

  # Return a list of ggplots for each individual gene
  myplots <- obsPlot(tidyInt,
                      plotByCol="GeneID",
                      groupCol = "group",
                      valueCol ="Log2CPM",
                      pointJitter = 0.1,
                      facet = FALSE)
  # Plot one from the list
  myplots[[2]]
}

}
