% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsPlot.R
\name{obsPlot}
\alias{obsPlot}
\title{Function obsPlot}
\usage{
obsPlot(
  DGEdata,
  plotType = "canvasXpress",
  countsMatrix = "counts",
  convertCounts = NULL,
  convert_geneLength,
  convert_log = FALSE,
  convert_normalize = "none",
  convert_prior.count = NULL,
  designTable = "design",
  group = "replicategroup",
  violinLayer = FALSE,
  showPoints = TRUE,
  xlab,
  ylab,
  title,
  boxColor = "deepskyblue3",
  facet = TRUE,
  facetRow,
  axisFree = TRUE
)
}
\arguments{
\item{plotType}{Can be canvasXpress or ggplot (default = canvasXpress)}

\item{designTable}{Name of the design table in the DGEObj from which the grouping column will be extracted. (default = design)}

\item{group}{Define the column name to group boxplots by (typically a replicate group column) (required)}

\item{violinLayer}{Adds a violin layer (default = FALSE)}

\item{showPoints}{Shows the datapoints on the plot only if boxLayer or violinLayer is enabled. (default = TRUE)}

\item{xlab}{X axis label (defaults to group column name if not specified)}

\item{ylab}{Y axis label (defaults to value column name if not specified)}

\item{title}{Plot title (optional)}

\item{boxColor}{Color for the boxplot layer (default = "deepskyblue3")}

\item{facet}{Specifies whether to facet (TRUE) or print individual plots
(FALSE)  (default = TRUE)}

\item{facetRow}{Explicitly set the number of rows for the facet plot.
Default behavior will automatically set the columns.}

\item{axisFree}{Specify same scale or independent scales for each subplot (default = TRUE;
Allowed values: TRUE and FALSE)}

\item{data}{A  DGEObject. The countsMatrix in the DGEObject is extracted to plot the data. (required)}

\item{boxLayer}{Adds a boxplot layer (default = TRUE)}

\item{showMean}{Show the mean of the points on the plot only if boxLayer is enabled. (default = TRUE)}

\item{boxTransparency}{Transparency for the box layer (default = 0.5)}

\item{violinColor}{Color for the violin layer (default = "yellow")}

\item{violinTransparency}{Transparency for the violin layer (default = 0.5)}

\item{labelAngle}{Angle to set the sample labels on the X axis. (default =  30; range = 0-90)}
}
\value{
Plot of type canvasXpress or ggplot. If Facet = TRUE (default) returns a faceted object. If
  facet = FALSE, returns a list of objects indexed
  by observation (gene) names.
}
\description{
Provides a summary plot (box/violin/points) for each observation (gene), showing data for each
experiment group. The plot can optionally include either one or both of the boxplot and violinplot layer
can additionally choose to display points and/or mean of all the points on the plot.  The boxLayer and the
violin layer can be customized to have the desired transparency, colors etc. By default, the violinLayer is
not displayed and only the boxplot, points and the mean on the plots can be seen.
}
\details{
Also, by default, the plots are faceted.  Faceting the plot can be turned off to return a list of individual
plots for each gene.

Input is a dataframe or matrix of intensity values with geneID row names and sample name column names,
e.g. from log2cpm. Group must also be specified as the plots are grouped by this attribute.
}
\examples{
\dontrun{

  # Faceted boxplot
  obsPlot(DGEobj,
          designTable = "design",
          group = "replicategroup")

  # Faceted violin plot
  obsPlot(DGEobj,
           violinLayer = TRUE,
           boxLayer = FALSE,
           designTable = "design",
           group = "replicategroup")

  # Return a list of plot for each individual gene
  myplots <- obsPlot(DGEobj,
                     designTable = "design",
                     group = "replicategroup"
                     facet = FALSE)
  # Plot one from the list
  myplots[[2]]
}

}
