% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispersion.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Plot edgeR dispersion from a DGEobj}
\usage{
plotDispersion(
  DGEdata,
  matrixType = "design",
  plotType = "canvasXpress",
  plotCategory = "dispersion",
  symbolSize = 6,
  symbolShape = "circle",
  symbolColor = "darkblue",
  symbolTransparency = 0.5,
  linefitColor = "red",
  lineFit = NULL,
  lineType = "solid",
  ...
)
}
\arguments{
\item{DGEdata}{DGEobj.}

\item{matrixType}{design or counts. If matrixType is design it will take the DGEList data in DGEobj and
if matrixType is counts it will take the counts data in DGEobj. (default = design)}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{plotCategory}{One of "dispersion" or "BCV" (default = "dispersion")}

\item{symbolSize}{(default = 6)}

\item{symbolShape}{Any supported shape for the points (default = "circle")
\url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types/}}

\item{symbolColor}{Fill color (default = "darkblue")}

\item{symbolTransparency}{Transparency for the points. Value ranges from 0 to 1. Smaller
indicate more transparency (default = 0.5)}

\item{linefitColor}{(default = "red")}

\item{lineFit}{(default = NULL) Any type supported by geom_smooth(if plotType is ggplot) or
one of glm, lm, loess, gam if plotType is canvasXpress. Loess is recommended.
  recommended.}

\item{lineType}{Any supported style for the fitLine  (default = solid).
\url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types/}}

\item{...}{Extra parameters to pass to edgeR::estimateDisp}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
Creates an edgeR dispersion plot for RNA-Seq QC purposes.  Takes a counts matrix
or DGEList for input.  Dispersion is plotted against AveLogCPM.  Optionally,
the plot can instead be Biological Coefficient of Variation (BCV is the square root of
dispersion) against AveLogCPM.
}
\examples{
\dontrun{
   # canvasxpress plot
   myCxplot <- plotDispersion(myDGEobj, matrixType = "counts")

   # ggplot
   myGgplot <- plotDispersion(myDGEobj, matrixType = "counts", plotType = "ggplot")
}

}
