% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispersion.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Plot edgeR dispersion from a DGEobj}
\usage{
plotDispersion(
  dgeObj,
  countsMatrix = TRUE,
  replicateGroupCol = "ReplicateGroup",
  plotType = "canvasXpress",
  plotCategory = "dispersion",
  lineFit = NULL,
  ...
)
}
\arguments{
\item{dgeObj}{DGEobj which has a counts matrix or a DGElist}

\item{countsMatrix}{If TRUE, uses the countsMatrix in DGEobj to construct the plot else DGElist will be used. (default = TRUE)}

\item{replicateGroupCol}{A singular value of class character and must be a column name in design object (default = "ReplicateGroup")}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{plotCategory}{One of "dispersion" or "BCV" (default = "dispersion")}

\item{lineFit}{(default = NULL) Any type supported by geom_smooth(if plotType is ggplot) or
one of glm, lm, loess, gam if plotType is canvasXpress. Loess is recommended.
  recommended.}

\item{...}{Extra parameters to pass to edgeR::estimateDisp}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
Creates an edgeR dispersion plot for RNA-Seq QC purposes.  Takes a DGeobj which has a counts matrix
or DGEList as input.  Dispersion is plotted against AveLogCPM.  Optionally,
the plot can instead be Biological Coefficient of Variation (BCV is the square root of
dispersion) against AveLogCPM.
}
\examples{
\dontrun{
   # canvasxpress plot
   myCxplot <- plotDispersion(myDGEobj)

   # ggplot
   myGgplot <- plotDispersion(myDGEobj, plotType = "ggplot")
}

}
