% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispersion.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Plot edgeR dispersion from a DGEobj}
\usage{
plotDispersion(
  DGEdata,
  designMatrix,
  plotType = "canvasXpress",
  plotValue = "dispersion",
  symbolSize = 1,
  symbolShape = 1,
  symbolColor = "darkblue",
  symbolFill = "darkblue",
  symbolAlpha = 0.3,
  linefitSize = 1,
  linefitColor = "yellow",
  lineFit = NULL,
  lineType = "solid",
  lineAlpha = 1,
  ...
)
}
\arguments{
\item{DGEdata}{Counts matrix or DGEList (Required)}

\item{designMatrix}{A design matrix created by stats::model.matrix (Required)}

\item{plotType}{Plot type must be canvasXpress or ggplot (Default to canvasXpress).}

\item{plotValue}{One of "dispersion" or "BCV" (Default = "dispersion")}

\item{symbolSize}{(Default = 1)}

\item{symbolShape}{see
http://www.cookbook-r.com/Graphs/Shapes_and_line_types/ (Default = 1)}

\item{symbolColor}{Default = "darkblue"}

\item{symbolFill}{Default = "darkblue"}

\item{symbolAlpha}{Transparency for the points. Value from 0 to 1. Smaller
indicate more transparency (Default = 0.3)}

\item{linefitSize}{(Default = 1)}

\item{linefitColor}{(Default = "yellow")}

\item{lineFit}{(Default = NULL) Any type supported by geom_smooth (Loess is
recommended).}

\item{lineType}{Any ggplot2 lineType supported by geom_smooth (Default = solid)}

\item{lineAlpha}{Alpha desired for geom_smooth line (Default = 1)}

\item{...}{Extra parameters to pass to edgeR::estimateDisp}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
Creates an edgeR dispersion plot for RNA-Seq QC purposes.  Takes a counts matrix
or DGEList for input.  Dispersion is plotted against AveLogCPM.  Optionally,
the plot can instead be Biological Coefficient of Variation (BCV is the square root of
dispersion) against AveLogCPM.
}
\examples{
\dontrun{
   # canvasxpress plot
   myCxplot <- plotDispersion(myDGElist)
   myCxplot <- plotDispersion(myDGEobj)

   # ggplot
   myGgplot <- plotDispersion(myDGElist, plotType = "ggplot")
   myGgplot <- plotDispersion(myDGEobj, plotType = "ggplot")
}

}
