% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispersion.R
\name{plotDispersion}
\alias{plotDispersion}
\title{Plot edgeR dispersion from a DGEobj}
\usage{
plotDispersion(
  DGEdata,
  designMatrix,
  plotType = "dispersion",
  symbolSize = 1,
  symbolShape = 1,
  symbolColor = "darkblue",
  symbolFill = "darkblue",
  symbolAlpha = 0.3,
  linefitSize = 1,
  linefitColor = "yellow",
  lineFit = NULL,
  lineType = 1,
  lineAlpha = 1,
  rugColor = NULL,
  rugAlpha = 0.02,
  ...
)
}
\arguments{
\item{DGEdata}{Counts matrix or DGEList (Required)}

\item{designMatrix}{A design matrix created by stats::model.matrix (Required)}

\item{plotType}{One of "dispersion" or "BCV" (Default = "dispersion")}

\item{symbolSize}{(Default = 1)}

\item{symbolShape}{see
http://www.cookbook-r.com/Graphs/Shapes_and_line_types/ (Default = 1)}

\item{symbolColor}{Default = "darkblue"}

\item{symbolFill}{Default = "darkblue"}

\item{symbolAlpha}{Transparency for the points. Value from 0 to 1. Smaller
indicate more transparency (Default = 0.3)}

\item{linefitSize}{(Default = 1)}

\item{linefitColor}{(Default = "yellow")}

\item{lineFit}{(Default = NULL) Any type supported by geom_smooth (Loess is
recommended).}

\item{lineType}{Any ggplot2 lineType supported by geom_smooth (Default = 1, solid)}

\item{lineAlpha}{Alpha desired for geom_smooth line (Default = 1)}

\item{rugColor}{(Default = NULL)  Set to Null to disable rug plots.}

\item{rugAlpha}{Transparency for the rug plots (Default = 0.02)}

\item{...}{Extra parameters to pass to edgeR::estimateDisp}
}
\value{
A ggplot object
}
\description{
Creates an edgeR dispersion plot for RNA-Seq QC purposes.  Takes a counts matrix
or DGEList for input.  Dispersion is plotted against AveLogCPM.  Optionally,
the plot can instead be Biological Coefficient of Variation (BCV is the square root of
dispersion) against AveLogCPM.
}
\examples{
\dontrun{
   myGgplot <- plotDispersion(myDGElist)
   myGgplot <- plotDispersion(myDGEobj)
}

}
