% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPValHist.R
\name{plotPvalHist}
\alias{plotPvalHist}
\title{Plot histogram analysis of p-value distributions}
\usage{
plotPvalHist(
  dgeObj,
  P.Val = "P.Value",
  plotType = "canvasXpress",
  facet = TRUE,
  binWidth = 0.02
)
}
\arguments{
\item{dgeObj}{DGEobj with topTables (required)}

\item{P.Val}{p-value column name in the topTables in DGEobj. Default="P.Value".}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = canvasXpress).}

\item{facet}{Set to FALSE to print individual plots instead of a faceted plot. (default = TRUE)}

\item{binWidth}{size of each bin.Value is always between 0 and 1. (default = 0.02)}
}
\value{
A canvasXpress or a ggplot2 object if facet = TRUE or a list of plots if facet = FALSE. (default = TRUE)
}
\description{
Generate a facet plot (or optionally individual plots) from a dataframe of
numbers. Intended to perform histogram analysis of p-value distributions,
but should be useful for any dataframe of numeric columns.
}
\examples{
\dontrun{
   # Plot to console (dgeObj is a DGEobj and P.Val is a name of topTable data in DGEobj.)
   dgeObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
   myplot <- plotPvalHist(dgeObj, P.Val = "P.Value")
   myplot <- plotPvalHist(dgeObj, P.Val = "P.Value", plotType = "ggplot")
}

}
