% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPValHist.R
\name{plotPvalHist}
\alias{plotPvalHist}
\title{Plot histogram analysis of p-value distributions}
\usage{
plotPvalHist(
  P.Val,
  plotType = "canvasXpress",
  facet = TRUE,
  binWidth = 0.02,
  alpha = 0.6,
  color = "dodgerblue3",
  fill = "dodgerblue3"
)
}
\arguments{
\item{P.Val}{A matrix or dataframe of numeric data; col = samples}

\item{plotType}{Plot type must be canvasXpress or ggplot (Default to canvasXpress).}

\item{facet}{Set to FALSE to print individual plots instead of a faceted plot. (Default = TRUE)}

\item{binWidth}{Range is always 0-1 for p-values. (Default = 0.02)}

\item{alpha}{Set the transparency. (Default = 0.6)}

\item{color}{Color for the histogram outline. (Default = "dodgerblue3")}

\item{fill}{Fill color for the histogram. (Default = "dodgerblue3")}
}
\value{
A ggplot2 object if facet = TRUE or a list of plots if facet = FALSE. (Default = TRUE)
}
\description{
Generate a facet plot (or optionally individual plots) from a dataframe of
numbers. Intended to perform histogram analysis of p-value distributions,
but should be useful for any dataframe of numeric columns.
}
\examples{
\dontrun{
   # Print to console using all defaults
   MyPvalMatrix <- extractCol(getType(myDGEobj, "topTable"), "P.Value")
   plotPvalHist(MyPvalMatrix)

   # Use some custom arguments
   myplot <- plotPvalHist(MyPValMatrix)
}

}
