% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Plot log intensity versus log ratio}
\usage{
profilePlot(
  contrastDF,
  plotType = "canvasXpress",
  logRatioCol = "logFC",
  logIntCol = "AveExpr",
  pvalCol = "P.Value",
  pthreshold = 0.01,
  geneSymLabels,
  geneSymCol,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  symbolSize = c(10, 4, 10),
  symbolShape = c("circle", "circle", "circle"),
  symbolColor = c("red3", "grey25", "deepskyblue4"),
  transparency = 0.5,
  sizeBySignificance = FALSE,
  referenceLine = "darkgoldenrod1",
  foldChangeLines = log2(1.5),
  refLineThickness = 1,
  lineFitType = "loess",
  lineFitColor = "goldenrod1",
  legendPosition = "right",
  footnote
)
}
\arguments{
\item{contrastDF}{dataframe with LogIntensity and LogRatio columns and optionally a p-value.}

\item{plotType}{Plot type must be canvasXpress or ggplot (default = "canvasXpress").}

\item{logRatioCol}{Name of the LogRatio column (default = "logFC")}

\item{logIntCol}{Name of the LogIntensity column (default = "AveExpr")}

\item{pvalCol}{Name of the p-value (default = "P.Value")}

\item{pthreshold}{Used to color points (default = 0.01)}

\item{geneSymLabels}{Gene Symbols to be labeled. For canvasXpress, these information will
be displayed for each data point.}

\item{geneSymCol}{Name of the gene symbol column in contrastDF.  The gene symbol is
not in topTable output by default so the user has to bind this column
to the dataframe in advance.  Then this column will be used to label
significantly changed points for ggplot type. For canvasXpress plot type,
gene information is displayed in the tooltip.}

\item{xlab}{X axis label (defaults = "LogIntensity column name")}

\item{ylab}{Y axis label (defaults = "LogRatio column name")}

\item{title}{Plot title (optional)}

\item{symbolSize}{Size of symbols for Up, no change, and Down (default = c(10, 4, 10));
Note: All three cannot be the same size. Decimal values are acceptable to help offset that
(e.g. 4, 4.1, 4.2).}

\item{symbolShape}{Shape of the symbols for Up, no change, and Down;
(default = c("circle", "circle", "circle").
See \url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{c(Up, NoChange, Down) (default = c("red3", "grey25", "deepskyblue4"))
See \url{http://research.stowers-institute.org/efg/R/Color/Chart}
Note: Colors cannot be duplicated.}

\item{transparency}{Controls the transparency of the plotted points (0-1; default = 0.5)}

\item{sizeBySignificance}{Set to TRUE to size points by the negative Log10 of the
Significance measure (default = FALSE)}

\item{referenceLine}{Color for an intercept = 0 horizontal reference line
(default = "darkgoldenrod1"; NULL disables)}

\item{foldChangeLines}{Position of reference horizontal lines for fold change
(default = log2(1.5); NULL disables)}

\item{refLineThickness}{Controls size of the horizontal reference line through geom_hline (default = 1)}

\item{lineFitType}{Enable a line fit through the data (default = "loess";
"lm" produces a linear fit. NULL disables)}

\item{lineFitColor}{Color for the fit line (default = "goldenrod1")}

\item{legendPosition}{One of "top", "bottom", "left", "right", "ne", "se", "nw", "sw", NULL.
top/bottom/left/right place the legend outside the figure.  ne/se/nw/sw place the figure
inside the figure. NULL disables the legend. Default = "right"}

\item{footnote}{Optional string placed right justified at bottom of plot.}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
A profile plot shows Log Intensity on the X axis and Log Ratio on the Y axis.
This function is intended to show the profile plot from a dataframe
in DGEobj design objects (i.e BDL_vs_SHAM).  Properly normalized data will generally be
centered around LogRatio = 0.
}
\details{
By default, the plot places "logFC" on the Y axis and "AveExpr" on the X
axis. By default, a reference horizontal line is shown at LogRatio = 0 on the
Y axis. Optionally, additional reference lines will be drawn at +/- a user
supplied LogRatio threshold. A Loess line fit is drawn through the actual
data. The points are color coded using the significance measure (i.e p-value)
supplied by the user. By default, the P.Value field is used
with a threshold of 0.01 to color code the points.

\strong{Data Structure for the input dataframe:}

The defaults are set for the contrasts on the DGEobj object (ie in BDL_vs_SHAM).
The columns named "logFC", "AveExpr", and "P.Value" are used by default to accommodate
the column names used in these dataframes.  Any other dataframe can be used with fold-change,
intensity, and significance measures, with appropriate arguments to define the column names
to use provided. By default, the column names will be used for the axis labels,
but can be overridden with xlab and ylab arguments.

A significance measure (which defaults to P.Value <= 0.01) is used to color code genes that
are significantly increased or decreased.

Sensible defaults are chosen for symbols (Size, Shape, Color, and Fill), but they can be
adjusted through the use of optional arguments. A length of 3 is
required for these arguments which applies the attributes in this order:
Increased, NoChange, Decreased.
}
\examples{
\dontrun{
   # Get DGEObj
   t_obj1 <- readRDS(system.file("exampleObj.RDS", package = "DGEobj", mustWork = TRUE))
   # Get Contrast Data to plot
   contrastDF <- t_obj1$BDL_vs_Sham
   myPlot <- profilePlot(contrastDF, title = "BDL_vs_Sham")

   # Some options with a custom datafile
   myPlot <- profilePlot(contrastDF,
                         pthreshold = 0.1,
                         title = "BDL_vs_Sham",
                         referenceLine = "blue",
                         legendPosition = "ne")
}

}
