% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create volcano plot}
\usage{
volcanoPlot(
  contrastDF,
  plotType = "canvasXpress",
  logRatioCol = "logFC",
  logIntCol = "AveExpr",
  pvalCol = "P.Value",
  pthreshold = 0.01,
  geneSymLabels,
  geneSymCol,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  symbolSize = c(10, 4, 10),
  symbolShape = c("circle", "circle", "circle"),
  symbolColor = c("red3", "grey25", "deepskyblue4"),
  sizeByIntensity = TRUE,
  transparency = 0.5,
  referenceLine = NULL,
  foldChangeLines = log2(1.5),
  refLineThickness = 2,
  legendPosition = "right",
  footnote
)
}
\arguments{
\item{contrastDF}{A dataframe with LogRatio and LogIntensity columns and optionally a
p-value or FDR column (typically a topTable dataframe).}

\item{plotType}{Plot type must be canvasXpress or ggplot (Default to canvasXpress).}

\item{logRatioCol}{Name of the LogRatio column (Default = "logFC")}

\item{logIntCol}{Name of the LogIntensity column (Default = "AveExpr")}

\item{pvalCol}{Name of the p-value or FDR column (Default = "P.Value")}

\item{pthreshold}{Used to color points (Default = 0.01)}

\item{geneSymLabels}{A character vector of gene to label (must be the name space of the column
specified by geneSymCol)}

\item{geneSymCol}{Name of the gene symbol column in contrastDF.  The gene symbol is
not in topTable output by default so the user has to bind this column
to the dataframe in advance.  This column will be used to label
significantly changed points.}

\item{xlab}{X axis label (Default is the LogIntensity column name)}

\item{ylab}{Y axis label (Default is the LogRatio column name)}

\item{title}{Plot title (optional)}

\item{symbolSize}{Size of symbols for Up, no change, and Down. default = c(10, 4, 10);
Note: All three cannot be the same size. Decimal values are acceptable to help offset that
      (e.g. 4, 4.1, 4.2).}

\item{symbolShape}{Shape of the symbols for Up, no change, and Down; Default =
c("circle", "circle", "circle"); Note: The same symbol shape cannot
be selected for all three symbols. See
\url{http://www.cookbook-r.com/Graphs/Shapes_and_line_types}}

\item{symbolColor}{Set the fill color for the symbols. Note only symbols 21-25
are fillable. This will have no effect on other symbols. Default =
c("red3", "grey25", "deepskyblue4") Note: Colors cannot be duplicated.}

\item{sizeByIntensity}{If TRUE, creates a column to support sizeByIntensity. (Default = TRUE)}

\item{transparency}{Controls the transparency of the plotted points (range: 0-1;
default = 0.5)}

\item{foldChangeLines}{Position of reference vertical lines for fold change
(Default = log2(1.5); NULL disables)}

\item{refLineThickness}{Set the thickness for all reference lines (Default = 2).}

\item{legendPosition}{One of "top", "bottom", "left", "right", "ne", "se",
"nw", "sw", NULL. top/bottom/left/right place the legend outside the
figure.  ne/se/nw/sw place the figure inside the figure. NULL disables the
legend. Default = "right"}

\item{footnote}{Optional string placed right justified at bottom of plot.}
}
\value{
canvasxpress or ggplot object based on plotType selection
}
\description{
A volcano plot shows Log Ratio data on the X axis and Negative Log P-values (NLP) on the
Y axis. This function is intended to show the volcano plot from a dataframe
created by topTable or topTreat. Properly normalized data will generally be
centered around LogRatio = 0.
}
\details{
By default, the plot places "logFC" on the X axis and Log10 of the "P.Value" on the Y axis.
By default, a reference vertical line is drawn at LogRatio = 0 on the X axis.
Optionally, additional reference lines will be drawn at +/- a user supplied Log Ratio threshold.
The points are color coded using both the significance and fold-change thresholds supplied by the user.
By default, the P.Value field is used with a threshold of 0.01 to color code the points and fold-change
threshold of +/- 1.5X.

\strong{Data Structure for the input dataframe:}

The defaults are set for dataframes produced by topTable and topTreat.  The columns named "logFC"
and "P.Value" are used by default to accommodate the column
names used in topTable/topTreat dataframes.  Any other dataframe
can be used with fold-change, intensity, and significance measures, with appropriate
arguments to define the column names to use provided. By default, the
column names will be used for the axis labels, but can be overridden with xlab and ylab arguments.

A significance measure (which defaults to P.Value <= 0.01) and LogRatio
threshold are used to color code genes that are significantly increased or decreased.
Use the appropriate arguments to use an FDR measure instead of p-value.

Sensible defaults are chosen for symbols (Size, Shape,  and Color), but they can be
adjusted through the use of optional arguments. A length of 3 is
required for these arguments which applies the attributes in this order:
Increased, NoChange, Decreased.
}
\examples{
\dontrun{
   # Simple plot with custom title (contrastDF is a topTable dataframe)
   myPlot <- volcanoPlot(contrastDF, title = "Plot Title")

   # Some options with a custom datafile
   myPlot <- volcanoPlot(contrastDF,
                         pthreshold = 0.1,
                         logRatioCol = "logFC",
                         logIntCol = "AveExpr",
                         pvalCol = "P.Value",
                         xlab = "Log2 Ratio", ylab = "negLog10p",
                         title = "Volcano Plot Title",
                         referenceLine = "blue",
                         legendPosition = "right")
}

}
