% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create volcano plot}
\usage{
volcanoPlot(
  dgeObj,
  contrast,
  plotType = "canvasXpress",
  logRatioCol = "logFC",
  logIntCol = "AveExpr",
  pvalCol = "P.Value",
  pthreshold = 0.01,
  geneNameCol,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  sizeByIntensity = TRUE,
  pthresholdLine = NULL,
  foldChangeThreshold = 1.5
)
}
\arguments{
\item{dgeObj}{DGEobj with a topTable dataframe (required)}

\item{contrast}{Name of the contrast.}

\item{plotType}{Plot type must be canvasXpress or ggplot. (default = canvasXpress)}

\item{logRatioCol}{Name of the LogRatio column. (default = "logFC")}

\item{logIntCol}{Name of the LogIntensity column. (default = "AveExpr")}

\item{pvalCol}{Name of the p-value or FDR column. (default = "P.Value")}

\item{pthreshold}{Used to color points. (default = 0.01)}

\item{geneNameCol}{geneName column in geneData from DGEobj. This column will be used to label significantly changed points.}

\item{xlab}{X axis label. (default is the LogIntensity column name)}

\item{ylab}{Y axis label. (default is the LogRatio column name)}

\item{title}{Plot title (optional).}

\item{sizeByIntensity}{If TRUE, creates a column to support sizeByIntensity. (default = TRUE)}

\item{pthresholdLine}{Color for a horizontal line at the p-threshold. (default = NULL (disabled))}

\item{foldChangeThreshold}{Position of reference vertical lines for fold change. (default = 1.5)}
}
\value{
canvasxpress or ggplot object based on plotType selection.
}
\description{
A volcano plot shows Log Ratio data on the X axis and Negative Log P-values (NLP) on the
Y axis. This function is intended to show the volcano plot created from a
topTable dataframe in a DGEobj. Properly normalized data will generally be
centered around LogRatio = 0.
}
\details{
By default, the plot places "logFC" on the X axis and Log10 of the "P.Value" on the Y axis.
By default, a reference vertical line is drawn at LogRatio = 0 on the X axis.
Optionally, additional reference lines will be drawn at +/- a user supplied Log Ratio threshold.
The points are color coded using both the significance and fold-change thresholds supplied by the user.
By default, the P.Value field is used with a threshold of 0.01 to color code the points and fold-change
threshold of +/- 1.5X.

\strong{Data Structure for the input DGEobj:}

A contrast needs to be specified along with the DGEobj. The top table is extracted from the DGEobj for this contrast.
The columns named "logFC" and "P.Value" in the topTable are used by default to generate the volcano plot.
By default, the column names will be used for the axis labels, but can be overridden with xlab and ylab arguments.

A significance measure (which defaults to P.Value <= 0.01) and LogRatio
threshold are used to color code genes that are significantly increased or decreased.
Use the appropriate arguments to use an FDR measure instead of p-value.
}
\examples{
\dontrun{
   # Simple plot with custom title (dgeObj is a name of DGEobj and contrast is a name of topTable dataframe)
   contrast <- names(DGEobj::getItems(dgeObj, "topTable"))[1]
   myPlot <- volcanoPlot(dgeObj,
                         contrast,
                         title = "Plot Title")

   # Some options with a custom datafile
   myPlot <- volcanoPlot(dgeObj,
                         contrast,
                         pthreshold     = 0.1,
                         logRatioCol    = "logFC",
                         logIntCol      = "AveExpr",
                         pvalCol        = "P.Value",
                         xlab           = "logFC",
                         ylab           = "negLog10p",
                         title          = "Volcano Plot Title",
                         pthresholdLine = "blue")

   myPlot <- volcanoPlot(dgeObj,
                         contrast,
                         pthreshold     = 0.1,
                         logRatioCol    = "logFC",
                         logIntCol      = "AveExpr",
                         pvalCol        = "P.Value",
                         xlab           = "logFC",
                         ylab           = "negLog10p",
                         title          = "Volcano Plot Title",
                         pthresholdLine = "blue",
                         plotType       = "ggplot")
}

}
